% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_resource.R
\name{add_resource}
\alias{add_resource}
\title{Add ETN data to a Frictionless Data Package}
\usage{
add_resource(package, resource_name, data)
}
\arguments{
\item{package}{Data Package object, as returned by \code{\link[frictionless:read_package]{read_package()}} or
\code{\link[frictionless:create_package]{create_package()}}.}

\item{resource_name}{Name of the Data Resource.}

\item{data}{bla}
}
\value{
Provided \code{package} with one additional resource.
}
\description{
Adds ETN data (\code{animals}, \code{deployments}, \code{detections}, \code{receivers}, \code{tags})
as a  Data Resource to a Frictionless Data Package.
The function extends \code{\link[frictionless:add_resource]{frictionless::add_resource()}}.
}
\seealso{
Other frictionless functions: 
\code{\link{reexports}}
}
\concept{frictionless functions}
