% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_date_time}
\alias{check_date_time}
\title{Check if the string input can be converted to a date}
\usage{
check_date_time(date_time, date_name = "start_date")
}
\arguments{
\item{date_time}{A character representation of a date.}

\item{date_name}{Informative description to user about type of date.}
}
\value{
FALSE | character
}
\description{
Returns FALSE or the cleaned character version of the date
(acknowledgments to micstr/isdate.R).
}
\examples{
\dontrun{
check_date_time("1985-11-21")
check_date_time("1985-11")
check_date_time("1985")
check_date_time("1985-04-31") # invalid date
check_date_time("01-03-1973") # invalid format
}
}
\keyword{internal}
