% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_value}
\alias{check_value}
\title{Check input value against list of provided values}
\usage{
check_value(arg, options = NULL, arg_name)
}
\arguments{
\item{arg}{Character. The input argument provided by the user.}

\item{options}{Character vector of valid inputs for the argument.}

\item{arg_name}{Character. The name of the argument used in the function to
test.}
}
\value{
If no error, \code{TRUE}.
}
\description{
Will return error message if an input value cannot be found in list of provided
values. NULL values are allowed.
}
\examples{
\dontrun{
# Valid inputs for tag_type
check_value("acoustic", c("acoustic", "archival"), "tag_type")
check_value(NULL, c("acoustic", "archival"), "tag_type")
check_value(c("acoustic", "archival"), c("acoustic", "archival"), "tag_type")

# Invalid inputs for tag_type
check_value("ddsf", c("acoustic", "archival"), "tag_type")
}
}
\keyword{internal}
