% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{collapse_transformer}
\alias{collapse_transformer}
\title{Support function for printing option help message}
\usage{
collapse_transformer(regex = "[*]$", ...)
}
\arguments{
\item{regex}{A regular expression to parse}

\item{...}{Additional arguments passed to the collapse}
}
\description{
Support function for printing option help message
}
