% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_etn.R
\name{connect_to_etn}
\alias{connect_to_etn}
\title{Connect to the ETN database}
\usage{
connect_to_etn(...)
}
\arguments{
\item{...}{Any arguments passed to this function are ignored.}
}
\value{
This function is no longer in use, and returns \code{NULL} invisibly.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{connect_to_etn()} is deprecated. It is no longer supported or needed. All
connections to the ETN database are now handled automatically when you use a
function. If your credentials are not stored in the system environment, you
will be prompted to enter them.
}
\examples{
\dontrun{
# This will result in a deprecation warning!
my_connection <- connect_to_etn()
}
}
\concept{connect functions}
