% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_acoustic_detections.R
\name{count_acoustic_detections2}
\alias{count_acoustic_detections2}
\title{Count acoustic detections}
\usage{
count_acoustic_detections2(
  credentials = list(username = Sys.getenv("ETN_USER"), password = Sys.getenv("ETN_PWD")),
  start_date = NULL,
  end_date = NULL,
  tag_serial_number = NULL,
  acoustic_tag_id = NULL,
  animal_project_code = NULL,
  scientific_name = NULL,
  acoustic_project_code = NULL,
  deployment_id = NULL,
  receiver_id = NULL,
  station_name = NULL
)
}
\description{
Count acoustic detections
}
