% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_detections.R
\name{deployment_station_names}
\alias{deployment_station_names}
\title{Support function to get unique set of deployment_station_name}
\usage{
deployment_station_names(connection)
}
\arguments{
\item{connection}{A valid connection to ETN database.}
}
\value{
A vector of all deployment_station_names present in vliz.deployments.
}
\description{
Get unique deployment_station_name
}
