% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_acoustic_dataset.R
\name{download_acoustic_dataset}
\alias{download_acoustic_dataset}
\title{Download acoustic data package}
\usage{
download_acoustic_dataset(
  connection = con,
  animal_project_code,
  scientific_name = NULL,
  directory = animal_project_code
)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}

\item{animal_project_code}{Character. Animal project you want to download
data for. Required.}

\item{scientific_name}{Character (vector). One or more scientific names.
Defaults to no all (all scientific names, include "Sync tag", etc.).}

\item{directory}{Character. Relative path to local download directory.
Defaults to creating a directory named after animal project code. Existing
files of the same name will be overwritten.}
}
\description{
Download all acoustic data related to an \strong{animal project} as a data
package that can be deposited in a research data repository. Includes option
to filter on scientific names.
}
\details{
The data are downloaded as a
\strong{\href{https://specs.frictionlessdata.io/data-package/}{Frictionless Data Package}}
containing:\tabular{ll}{
   file \tab description \cr
   \code{animals.csv} \tab Animals related to an \code{animal_project_code}, optionally filtered on \code{scientific_name}(s), as returned by \code{get_animals()}. \cr
   \code{tags.csv} \tab Tags associated with the selected animals, as returned by \code{get_tags()}. \cr
   \code{detections.csv} \tab Acoustic detections for the selected animals, as returned by \code{get_acoustic_detections()}. \cr
   \code{deployments.csv} \tab Acoustic deployments for the \code{acoustic_project_code}(s) found in detections, as returned by \code{get_acoustic_deployments()}. This allows users to see when receivers were deployed, even if these did not detect the selected animals. \cr
   \code{receivers.csv} \tab Acoustic receivers for the selected deployments, as returned by \code{get_acoustic_receivers()}. \cr
   \code{datapackage.json} \tab A \href{https://specs.frictionlessdata.io/table-schema/}{Frictionless Table Schema} metadata file describing the fields and relations of the above csv files. This file is copied from \href{https://github.com/inbo/etn/blob/master/inst/assets/datapackage.json}{here} and can be used to validate the data package. \cr
}


The function will report the number of records per csv file, as well as the
included scientific names and acoustic projects. Warnings will be raised for:
\itemize{
\item Animals with multiple tags
\item Tags associated with multiple animals
\item Deployments without acoustic project: these deployments will not be listed
in \code{deployments.csv} and will therefore raise a foreign key validation
error.
\item Duplicate detections: detections with the duplicate \code{detection_id}. These
are removed by the function in \code{detections.csv}.
}

\strong{Important}: The data are downloaded \emph{as is} from the database, i.e. no
quality or consistency checks are performed by this function. We therefore
recommend to verify the data before publication. A consistency check can be
performed by validation tools of the Frictionless Framework, e.g.
\verb{frictionless validate datapackage.json} on the command line using
\href{https://github.com/frictionlessdata/frictionless-py}{frictionless-py}.
}
\examples{
\dontrun{
# Set default connection variable
con <- connect_to_etn()

# Download data for the 2012_leopoldkanaal animal project (all scientific names)
download_acoustic_dataset(animal_project_code = "2012_leopoldkanaal")
#> Downloading data to directory `2012_leopoldkanaal`:
#> * (1/6): downloading animals.csv
#> * (2/6): downloading tags.csv
#> * (3/6): downloading detections.csv
#> * (4/6): downloading deployments.csv
#> * (5/6): downloading receivers.csv
#> * (6/6): adding datapackage.json as file metadata
#>
#> Summary statistics for dataset `2012_leopoldkanaal`:
#> * number of animals:           104
#> * number of tags:              103
#> * number of detections:        2215243
#> * number of deployments:       1968
#> * number of receivers:         454
#> * first date of detection:     2012-07-04
#> * last date of detection:      2021-09-02
#> * included scientific names:   Anguilla anguilla
#> * included acoustic projects:  albert, Apelafico, bpns, JJ_Belwind, leopold, MOBEIA, pc4c, SPAWNSEIS, ws2, zeeschelde
#>
#> Warning message:
#> In download_acoustic_dataset(animal_project_code = "2012_leopoldkanaal") :
#> Found tags associated with multiple animals: 1145373
}
}
