% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\title{Download data for publication}
\usage{
download_dataset(
  connection = con,
  animal_project_code = NULL,
  directory = animal_project_code,
  scientific_name = NULL
)
}
\arguments{
\item{connection}{A valid connection to the ETN database.}

\item{animal_project_code}{(string) Animal project to download data from.}

\item{directory}{(string) Path to local download directory.}

\item{scientific_name}{(string) One or more scientific names to filter on.}
}
\value{
...
}
\description{
...
}
\examples{
\dontrun{
...
}
}
