% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etn.R
\docType{package}
\name{etn-package}
\alias{etn}
\alias{etn-package}
\title{etn: Access Data from the European Tracking Network}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access data from the European Tracking Network ('ETN', \url{https://www.lifewatch.be/etn/}). With etn you can query metadata (animals, tags, deployments, receivers, projects) and data (acoustic detections, sensor data) from the 'ETN' database and use these in your analyses. Data access requires user credentials and is subject to a data policy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/inbo/etn}
  \item \url{https://inbo.github.io/etn/}
  \item Report bugs at \url{https://github.com/inbo/etn/issues}
}

}
\author{
\strong{Maintainer}: Pieter Huybrechts \email{pieter.huybrechts@inbo.be} (\href{https://orcid.org/0000-0002-6658-6062}{ORCID})

Authors:
\itemize{
  \item Peter Desmet \email{peter.desmet@inbo.be} (\href{https://orcid.org/0000-0002-8442-8025}{ORCID})
  \item Damiano Oldoni \email{damiano.oldoni@inbo.be} (\href{https://orcid.org/0000-0003-3445-7562}{ORCID})
  \item Stijn Van Hoey \email{stijnvanhoey@gmail.com} (\href{https://orcid.org/0000-0001-6413-3185}{ORCID})
}

Other contributors:
\itemize{
  \item Research Institute for Nature and Forest (INBO) (\href{https://ror.org/00j54wy13}{ROR}) [copyright holder]
  \item Research Foundation - Flanders (https://lifewatch.be) [funder]
}

}
\keyword{internal}
