% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_etn.R
\docType{data}
\name{ETN_SERVER}
\alias{ETN_SERVER}
\title{ETN database informations}
\format{An object of class \code{character} of length 1.}
\usage{
ETN_SERVER
}
\description{
ETN database informations
}
\keyword{datasets}
