% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\name{example_dataset}
\alias{example_dataset}
\title{Read an example dataset}
\usage{
example_dataset(dataset = "2014_DEMER")
}
\arguments{
\item{dataset}{Name of the example dataset to load. Defaults to the only
available dataset: \code{"2014_DEMER"}.}
}
\value{
Frictionless data package.
}
\description{
Reads an example dataset, formatted as a \href{https://specs.frictionlessdata.io/data-package/}{Frictionless Data Package}.
}
\section{2014_DEMER}{


\verb{2014_DEMER} is a \strong{river acoustic telemetry} dataset. It contains 66 tagged
animals across four species. Over 235,000 detections were observed between
2014 and 2019 by acoustic receivers deployed in Belgian rivers.

Data are deposited at \url{https://doi.org/10.14284/432} and can be downloaded
from the ETN database with
\code{download_acoustic_dataset(animal_project_code = "2014_DEMER")}.
}

\examples{
example_dataset()
}
\concept{sample data}
