% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acoustic_citations.R
\name{get_acoustic_citations}
\alias{get_acoustic_citations}
\title{Get Acoustic Citations}
\usage{
get_acoustic_citations(acoustic_project_code = NULL)
}
\arguments{
\item{acoustic_project_code}{Character (vector). One or more acoustic
project codes. Case-insensitive.}
}
\value{
(Invisible) A named vector citations for the specified acoustic
projects. And prints to console the formatted citations as shown on IMIS.
}
\description{
This function returns the the citations for all acoustic (= network) projects
included in a dataset. These citations will be retrieved from
\href{https://www.vliz.be/nl/imis}{IMIS}.
}
\examples{
get_acoustic_citations("2004_Gudena")
}
