% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acoustic_deployments.R
\name{get_acoustic_deployments}
\alias{get_acoustic_deployments}
\title{Get acoustic deployment data}
\usage{
get_acoustic_deployments(
  connection,
  deployment_id = NULL,
  receiver_id = NULL,
  acoustic_project_code = NULL,
  station_name = NULL,
  open_only = FALSE
)
}
\arguments{
\item{connection}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A connection to the ETN
database. This argument is no longer used. You will be prompted for
credentials instead.}

\item{deployment_id}{Integer (vector). One or more deployment identifiers.}

\item{receiver_id}{Character (vector). One or more receiver identifiers.}

\item{acoustic_project_code}{Character (vector). One or more acoustic
project codes. Case-insensitive.}

\item{station_name}{Character (vector). One or more deployment station
names.}

\item{open_only}{Logical. Restrict deployments to those that are currently
open (i.e. no end date defined). Defaults to \code{FALSE}.}
}
\value{
A tibble with acoustic deployment data, sorted by
\code{acoustic_project_code}, \code{station_name} and \code{deploy_date_time}.
}
\description{
Get data for deployments of acoustic receivers, with options to filter
results.
}
\examples{
\dontshow{if (etn:::credentials_are_set()) withAutoprint(\{ # examplesIf}
# Get all acoustic deployments
get_acoustic_deployments()

# Get specific acoustic deployment
get_acoustic_deployments(deployment_id = 1437)

# Get acoustic deployments for a specific receiver
get_acoustic_deployments(receiver_id = "VR2W-124070")

# Get open acoustic deployments for a specific receiver
get_acoustic_deployments(receiver_id = "VR2W-124070", open_only = TRUE)

# Get acoustic deployments for a specific acoustic project
get_acoustic_deployments(acoustic_project_code = "demer")

# Get acoustic deployments for two specific stations
get_acoustic_deployments(station_name = c("de-9", "de-10"))
\dontshow{\}) # examplesIf}
}
\seealso{
Other access functions:
\code{\link[=get_acoustic_deployment_logs]{get_acoustic_deployment_logs()}},
\code{\link[=get_acoustic_detections]{get_acoustic_detections()}},
\code{\link[=get_acoustic_projects]{get_acoustic_projects()}},
\code{\link[=get_acoustic_receivers]{get_acoustic_receivers()}},
\code{\link[=get_animal_projects]{get_animal_projects()}},
\code{\link[=get_animals]{get_animals()}},
\code{\link[=get_archival_data]{get_archival_data()}},
\code{\link[=get_bibliography]{get_bibliography()}},
\code{\link[=get_cpod_projects]{get_cpod_projects()}},
\code{\link[=get_package]{get_package()}},
\code{\link[=get_tags]{get_tags()}}
}
\concept{access functions}
