% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acoustic_tags.R
\name{get_acoustic_tags}
\alias{get_acoustic_tags}
\title{Get acoustic tag data}
\usage{
get_acoustic_tags(
  connection = con,
  tag_serial_number = NULL,
  acoustic_tag_id = NULL,
  include_ref_tags = FALSE
)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}

\item{tag_serial_number}{Character (vector). One or more tag serial numbers.}

\item{acoustic_tag_id}{Character (vector). One or more acoustic tag ids.}

\item{include_ref_tags}{Logical. Include reference tags. Defaults to
\code{FALSE}.}
}
\value{
A tibble with tags data, sorted by \code{tag_serial_number}. See also
\href{https://inbo.github.io/etn/articles/etn_fields.html}{field definitions}.
}
\description{
Get data for acoustic tags, with options to filter results. By default,
reference tags are excluded.
}
\examples{
\dontrun{
con <- connect_to_etn(your_username, your_password)

# Get all acoustic tags
get_tags(con)

# Get all acoustic tags, including reference tags
get_tags(con, include_ref_tags = TRUE)

# Get specific acoustic tags (will automatically set include_ref_tags = TRUE)
get_tags(con, tag_id = "A69-1303-65313")
get_tags(con, tag_id = c("A69-1601-1705", "A69-1601-1707"))
}
}
