% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archival_data.R
\name{get_archival_data_uuid}
\alias{get_archival_data_uuid}
\title{Fetch a table with UUID references to archival data files}
\usage{
get_archival_data_uuid(
  tag_serial_number = NULL,
  animal_id = NULL,
  animal_project_code = NULL
)
}
\arguments{
\item{tag_serial_number}{Character (vector). One or more tag serial numbers.}

\item{animal_id}{Integer (vector). One or more animal identifiers.}

\item{animal_project_code}{Character (vector). One or more animal project
codes. Case-insensitive.}
}
\value{
A tibble with a column \code{converted_archival_file_uuid} that contains
the UUID pointing to the archival data csv file. Other columns are:
\code{animal_id}, \code{tag_serial_number}, and \code{animal_project_code}.
}
\description{
Archival data is stored in csv files hosted on lifewatch.com. To fetch these
we must first query ETN for uuids that are needed to build the paths to read
these csv files.
}
\examples{
\dontshow{if (interactive() & credentials_are_set()) withAutoprint(\{ # examplesIf}
get_archival_data_uuid()
get_archival_data_uuid(tag_serial_number = "22035610")
\dontshow{\}) # examplesIf}
}
\concept{helper functions}
