% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archival_tags.R
\name{get_archival_tags}
\alias{get_archival_tags}
\title{Get archival tag data}
\usage{
get_archival_tags(
  connection = con,
  tag_serial_number = NULL,
  archival_tag_id = NULL
)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}

\item{tag_serial_number}{Character (vector). One or more tag serial numbers.}

\item{archival_tag_id}{Character (vector). One or more archival tag
identifiers.}
}
\value{
A tibble with tags data, sorted by \code{tag_serial_number}. See also
\href{https://inbo.github.io/etn/articles/etn_fields.html}{field definitions}.
}
\description{
Get data for archival tags, with options to filter results. Note that there
can be multiple records (\code{archival_tag_id}) per tag device
(\code{tag_serial_number}), often multiple types of sensors
}
\examples{
\dontrun{
# Set default connection variable
con <- connect_to_etn()

# Get all archival tags
get_archival_tags()

# Get specific archival tags
get_archival_tags(tag_serial_number = "1292638")
get_archival_tags(archival_tag_id = "3638")
get_archival_tags(archival_tag_id = c("3638", "3639"))
}
}
