% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpod_deployments.R
\name{get_cpod_deployments}
\alias{get_cpod_deployments}
\title{Get cpod deployment data}
\usage{
get_cpod_deployments(
  connection = con,
  deployment_id = NULL,
  receiver_id = NULL,
  cpod_project_code = NULL,
  station_name = NULL,
  open_only = FALSE
)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}

\item{deployment_id}{Integer (vector). One or more deployment identifiers.}

\item{receiver_id}{Character (vector). One or more receiver identifiers.}

\item{cpod_project_code}{Character (vector). One or more cpod
project codes. Case-insensitive.}

\item{station_name}{Character (vector). One or more deployment station
names.}

\item{open_only}{Logical. Restrict deployments to those that are currently
open (i.e. no end date defined). Defaults to \code{FALSE}.}
}
\value{
A tibble with cpod deployment data, sorted by
\code{cpod_project_code}, \code{station_name} and \code{deploy_date_time}. See also
\href{https://inbo.github.io/etn/articles/etn_fields.html}{field definitions}.
}
\description{
Get data for deployments of cpod receivers, with options to filter
results.
}
\examples{
# Set default connection variable
con <- connect_to_etn()

# Get all cpod deployments
get_cpod_deployments(con)

# Get specific cpod deployment
get_cpod_deployments(con, deployment_id = 1437)

# Get cpod deployments for a specific receiver
get_cpod_deployments(con, receiver_id = "VR2W-124070")

# Get open cpod deployments for a specific receiver
get_cpod_deployments(con, receiver_id = "VR2W-124070", open_only = TRUE)

# Get cpod deployments for a specific cpod project
get_cpod_deployments(con, cpod_project_code = "demer")

# Get cpod deployments for two specific stations
get_cpod_deployments(con, station_name = c("de-9", "de-10"))
}
