% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deployments.R
\name{get_deployments}
\alias{get_deployments}
\title{Get deployment metadata}
\usage{
get_deployments(
  connection,
  network_project = NULL,
  receiver_status = NULL,
  open_only = TRUE
)
}
\arguments{
\item{connection}{A valid connection to the ETN database.}

\item{network_project}{(string) One or more network projects.}

\item{open_only}{(logical) Restrict to deployments that are currently open (i.e. no end date defined).  Default:
\code{TRUE}.}
}
\value{
A tibble (tidyverse data.frame).
}
\description{
Get metadata for deployments, with options to filter on network project,
receiver status and/or open deployments.
}
\examples{
\dontrun{
con <- connect_to_etn(your_username, your_password)

# Get all open deployments
get_deployments(con)

# Get all deployments (including those with an end date)
get_deployments(con, open_only = FALSE)

# Get open deployments from specific network project(s)
get_deployments(con, network_project = c("thornton", "leopold"))

# Get open deployments with a specific receiver status
get_deployments(con, receiver_status = c("Broken", "Lost"))

# Get open deployments from a specific project from active receivers
get_deployments(con, network_project = "thornton", receiver_status = "Active")
}
}
