% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_detections_old.R
\name{get_detections_old}
\alias{get_detections_old}
\title{Get detections data}
\usage{
get_detections_old(
  connection = con,
  application_type = NULL,
  network_project_code = NULL,
  animal_project_code = NULL,
  start_date = NULL,
  end_date = NULL,
  station_name = NULL,
  tag_id = NULL,
  receiver_id = NULL,
  scientific_name = NULL,
  limit = FALSE
)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}

\item{application_type}{Character. \code{acoustic_telemetry} or \code{cpod}.}

\item{network_project_code}{Character (vector). One or more network
projects. Case-insensitive.}

\item{animal_project_code}{Character (vector). One or more animal projects.
Case-insensitive.}

\item{start_date}{Character. Date in ISO 8601 format, e.g. \code{2018-01-01}.
Date selection on month (e.g. \code{2018-03}) or year (e.g. \code{2018}) are
supported as well.}

\item{end_date}{Character. Date in ISO 8601 format, e.g. \code{2018-01-01}.
Date selection on month (e.g. 2018-03) or year (e.g. 2018) are
supported as well.}

\item{station_name}{Character (vector). One or more deployment station
names. Case-insensitive.}

\item{tag_id}{Character (vector). One or more tag identifiers.}

\item{receiver_id}{Character (vector). One or more receiver identifiers.}

\item{scientific_name}{Character (vector). One or more scientific names.}

\item{limit}{Logical. Limit the number of returned records to 100 (useful
for testing purposes). Defaults to \code{FALSE}.}
}
\value{
A tibble with detections data, sorted by \code{tag_id} and \code{date_time}.
See also
\href{https://inbo.github.io/etn/articles/etn_fields.html}{field definitions}.
}
\description{
Get detections data, with options to filter results. Use \code{limit} to limit
the number of returned records.
}
\examples{
\dontrun{
con <- connect_to_etn(your_username, your_password)

# Get detections filtered by start year, limited to 100 records by default
get_detections_old(con, start_date = "2017")

# Get detections for a specific application type, limited to 100 records
get_detections_old(con, application_type = "acoustic_telemetry")

# Get detections within a time frame for a specific animal project and
# network project
get_detections_old(
  con,
  animal_project_code = "phd_reubens",
  network_project_code = "thornton", start_date = "2011-01-28",
  end_date = "2011-02-01"
)

# Get detections for a specific animal project at specific stations
get_detections_old(
  con,
  animal_project_code = "phd_reubens",
  station_name = c("R03", "R05")
)

# Get detections for a specific tag
get_detections_old(
  con,
  tag_id = "A69-1303-65302"
)

# Get detections for a specific receiver during a specific time period
get_detections_old(
  con,
  receiver_id = "VR2W-122360",
  start_date = "2015-12-03",
  end_date = "2015-12-05"
)
# Get detections for a specific species during a given period
get_detections_old(
  con,
  scientific_name = "Anguilla anguilla",
  start_date = "2015-12-03",
  end_date = "2015-12-05"
)
}
}
