% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projects_old.R
\name{get_projects_old}
\alias{get_projects_old}
\title{Get project data}
\usage{
get_projects_old(
  connection = con,
  project_type = NULL,
  application_type = NULL
)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}

\item{project_type}{Character. \code{animal} or \code{network}.}

\item{application_type}{Character. \code{acoustic_telemetry} or \code{cpod}.}
}
\value{
A tibble with project data, sorted by \code{project_code}. See also
\href{https://inbo.github.io/etn/articles/etn_fields.html}{field definitions}.
}
\description{
Get data for projects, with options to filter results.
}
\examples{
\dontrun{
con <- connect_to_etn(your_username, your_password)

# Get all projects
get_projects_old(con)

# Get all animal projects
get_projects_old(con, project_type = "animal")

# Get all network projects
get_projects_old(con, project_type = "network")

# Get projects for a specific application type
get_projects_old(con, application_type = "cpod")
}
}
