% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_receiver_diagnostics.R
\name{get_receiver_diagnostics}
\alias{get_receiver_diagnostics}
\title{Get diagnostics information for a receiver for a deployment id.}
\usage{
get_receiver_diagnostics(
  connection = con,
  deployment_id,
  start_date = NULL,
  end_date = NULL,
  receiver_id = NULL,
  limit = FALSE
)
}
\arguments{
\item{connection}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A connection to the ETN
database. This argument is no longer used. You will be prompted for
credentials instead.}

\item{deployment_id}{Integer (vector). One or more deployment identifiers.}

\item{start_date}{Character. Start date (inclusive) in ISO 8601 format (
\code{yyyy-mm-dd}, \code{yyyy-mm} or \code{yyyy}).}

\item{end_date}{Character. End date (exclusive) in ISO 8601 format (
\code{yyyy-mm-dd}, \code{yyyy-mm} or \code{yyyy}).}

\item{receiver_id}{Character (vector). One or more receiver identifiers.}

\item{limit}{Logical. Limit the number of returned records to 100 (useful
for testing purposes). Defaults to \code{FALSE}.}
}
\value{
A tibble with receiver diagnostics data
}
\description{
Get diagnostics information for a receiver for a deployment id.
}
