% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transmitters.R
\name{get_transmitters}
\alias{get_transmitters}
\title{Get transmitter metadata}
\usage{
get_transmitters(connection, animal_project = NULL)
}
\arguments{
\item{connection}{A valid connection with the ETN database.}

\item{animal_project}{(string) One or more animal projects.}
}
\value{
A data.frame.
}
\description{
Get the metadata about the transmitter tags. At the moment, only
transmitters that can be linked to a projectcode are returned to the user.
}
\examples{
\dontrun{
con <- connect_to_etn(your_username, your_password)

# Get the metadata of all transmitter tags
get_transmitters(con)

# Get the metadata of the tags linked to specific project(s)
get_transmitters(con, animal_project = "phd_reubens")
get_transmitters(con, animal_project = c("phd_reubens", "2012_leopoldkanaal"))
}
}
