% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_val}
\alias{get_val}
\title{Retrieve the result of a function called to the opencpu api}
\usage{
get_val(temp_key, api_domain = "https://opencpu.lifewatch.be")
}
\arguments{
\item{temp_key}{the temp key returned from the POST request to the API}
}
\value{
the uncompressed object resulting form a GET request to the API
}
\description{
Loading the evaluated object into the current environment, to be used
internally in functions calling the opencpu api service to convert a response
object included in the response from a post request, to the corresponding
objects resulting from the original call.
}
\examples{
\dontrun{etn:::extract_temp_key(response) \%>\% get_val()}

# using the opencpu test instance
api_url <- "https://cloud.opencpu.org/ocpu/library/stats/R/rnorm"
httr::POST(api_url,body = list(n = 10,mean = 5)) \%>\%
    extract_temp_key() \%>\%
    get_val(api_domain = "https://cloud.opencpu.org/ocpu")
 )
}
