% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_animal_ids.R
\name{list_animal_ids}
\alias{list_animal_ids}
\title{List all available animal ids}
\usage{
list_animal_ids(connection = con)
}
\arguments{
\item{connection}{A valid connection to the ETN database.}
}
\value{
A vector of all unique \code{animal_id} present in \code{animals_view2}.
}
\description{
List all available animal ids
}
