% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_animal_ids_api.R
\name{list_animal_ids_api}
\alias{list_animal_ids_api}
\title{List all available animal ids}
\usage{
list_animal_ids_api(credentials = get_credentials())
}
\arguments{
\item{credentials}{Login credentials to the ETN database, as created by
\code{get_credentials()}}
}
\value{
A vector of all unique \code{id_pk} present in \code{common.animal_release}.
}
\description{
List all available animal ids
}
