% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_animal_project_codes.R
\name{list_animal_project_codes}
\alias{list_animal_project_codes}
\title{List all available animal project codes}
\usage{
list_animal_project_codes(connection)
}
\arguments{
\item{connection}{A valid connection to the ETN database.}
}
\value{
A vector of all unique \code{project_code} of \code{type="animal"} present in \code{projects}.
}
\description{
List all available animal project codes
}
