% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_archival_tag_ids.R
\name{list_archival_tag_ids}
\alias{list_archival_tag_ids}
\title{List all available archival tag ids}
\usage{
list_archival_tag_ids(connection = con)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}
}
\value{
A vector of all unique \code{id_code} present in \code{archive.sensor}.
}
\description{
List all available archival tag ids
}
