% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_cpod_project_codes.R
\name{list_cpod_project_codes}
\alias{list_cpod_project_codes}
\title{List all available cpod project codes}
\usage{
list_cpod_project_codes(connection, api = TRUE)
}
\arguments{
\item{connection}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A connection to the ETN
database. This argument is no longer used. You will be prompted for
credentials instead.}

\item{api}{Logical. If \code{TRUE}, the function will use the API to access ETN.}
}
\value{
A vector of all unique \code{project_code} of \code{type = "cpod"} in
\code{project.sql}.
}
\description{
List all available cpod project codes
}
