% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_deployment_ids.R
\name{list_deployment_ids}
\alias{list_deployment_ids}
\title{List all available receiver ids}
\usage{
list_deployment_ids(connection, api = TRUE)
}
\arguments{
\item{connection}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A connection to the ETN
database. This argument is no longer used. You will be prompted for
credentials instead.}

\item{api}{Logical. If \code{TRUE}, the function will use the API to access ETN.}
}
\value{
A vector of all unique \code{id_pk} present in \code{acoustic.deployments}.
}
\description{
List all available receiver ids
}
