% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_network_project_codes.R
\name{list_network_project_codes}
\alias{list_network_project_codes}
\title{List all available network project codes}
\usage{
list_network_project_codes(connection)
}
\arguments{
\item{connection}{A valid connection to the ETN database.}
}
\value{
A vector of all unique \code{project_code} of \code{project_type="network"} present in \code{projects}.
}
\description{
List all available network project codes
}
