% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_receiver_ids.R
\name{list_receiver_ids}
\alias{list_receiver_ids}
\title{List all available receiver ids}
\usage{
list_receiver_ids(connection)
}
\arguments{
\item{connection}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A connection to the ETN
database. This argument is no longer used. You will be prompted for
credentials instead.}
}
\value{
A vector of all unique \code{receiver} present in \code{acoustic.receivers}.
}
\description{
List all available receiver ids
}
\examples{
\dontshow{if (etn:::credentials_are_set()) withAutoprint(\{ # examplesIf}
list_receiver_ids()
\dontshow{\}) # examplesIf}
}
\seealso{
Other list functions: 
\code{\link{list_acoustic_project_codes}()},
\code{\link{list_acoustic_tag_ids}()},
\code{\link{list_animal_ids}()},
\code{\link{list_animal_project_codes}()},
\code{\link{list_cpod_project_codes}()},
\code{\link{list_deployment_ids}()},
\code{\link{list_scientific_names}()},
\code{\link{list_station_names}()},
\code{\link{list_tag_serial_numbers}()},
\code{\link{list_values}()}
}
\concept{list functions}
