% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_scientific_names.R
\name{list_scientific_names}
\alias{list_scientific_names}
\title{List all available scientific names}
\usage{
list_scientific_names(connection)
}
\arguments{
\item{connection}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A connection to the ETN
database. This argument is no longer used. You will be prompted for
credentials instead.}
}
\value{
A vector of all unique \code{scientific_name} present in
\code{common.animal_release}.
}
\description{
List all available scientific names
}
\examples{
\dontshow{if (etn:::credentials_are_set()) withAutoprint(\{ # examplesIf}
list_scientific_names()
\dontshow{\}) # examplesIf}
}
\seealso{
Other list functions: 
\code{\link{list_acoustic_project_codes}()},
\code{\link{list_acoustic_tag_ids}()},
\code{\link{list_animal_ids}()},
\code{\link{list_animal_project_codes}()},
\code{\link{list_cpod_project_codes}()},
\code{\link{list_deployment_ids}()},
\code{\link{list_receiver_ids}()},
\code{\link{list_station_names}()},
\code{\link{list_tag_serial_numbers}()},
\code{\link{list_values}()}
}
\concept{list functions}
