% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_station_names.R
\name{list_station_names}
\alias{list_station_names}
\title{List all available station names}
\usage{
list_station_names(connection)
}
\arguments{
\item{connection}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A connection to the ETN
database. This argument is no longer used. You will be prompted for
credentials instead.}
}
\value{
A vector of all unique \code{station_name} present in
\code{acoustic.deployments}.
}
\description{
List all available station names
}
\examples{
list_station_names()
}
