% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tag_ids.R
\name{list_tag_ids}
\alias{list_tag_ids}
\title{List all available tag ids}
\usage{
list_tag_ids(connection = con)
}
\arguments{
\item{connection}{A valid connection to the ETN database.}
}
\value{
A vector of all unique \code{tag_id} present in \code{tags_view2}.
}
\description{
List all available tag ids
}
