% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_values.R
\name{list_values}
\alias{list_values}
\title{List all unique values from a data.frame column}
\usage{
list_values(.data, column, split = ",")
}
\arguments{
\item{.data}{Data frame. Data.frame to select column from.}

\item{column}{Character or integer. Quoted or unqoted column name or column
position.}

\item{split}{Character (vector). Character or regular expression(s) passed
to \code{\link[=strsplit]{strsplit()}} to split column values before returning unique values.
Defaults to \verb{,}.}
}
\value{
A vector of the same type as the given column.
}
\description{
Get a vector with all unique values found in a given column of a data.frame.
Concatenated values (\verb{A,B}) in the column can be returned as single values
(\code{A} and \code{B}).
}
\examples{
\dontrun{
# List unique scientific_name from a dataframe containing animal information
df <- get_animals(animal_project_code = "2014_demer")
list_values(df, "scientific_name")

# Or using pipe and unquoted column name
df \%>\% list_values(scientific_name)

# Or using column position
df \%>\% list_values(df, 6)

# tag_serial_number contains comma-separated values
df <- get_animals(animal_id = 2825, 2827)
df$tag_serial_number

# Comma-seperated values are split by default
list_values(df, tag_serial_number)

# They are not split using another split expression (here ".")
list_values(df, tag_serial_number, split = "\\\\.")
}
}
