% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{network_projects}
\alias{network_projects}
\title{Get list of unique available network_project_code}
\usage{
network_projects(connection)
}
\arguments{
\item{connection}{A valid connection to the ETN database.}
}
\value{
A vector of all network_project_code present in vliz.projects.
}
\description{
Get list of unique available network_project_code
}
