% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frictionless-reexports.R
\name{read_package}
\alias{read_package}
\title{Read a Data Package descriptor file (\code{datapackage.json})}
\usage{
read_package(file = "datapackage.json")
}
\description{
Read a Data Package descriptor file (\code{datapackage.json})
}
\seealso{
Other frictionless functions: 
\code{\link{add_resource}()},
\code{\link{write_package}()}
}
\concept{frictionless functions}
