% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frictionless-reexports.R
\name{read_resource}
\alias{read_resource}
\title{Read data from a Data Resource into a tibble data frame}
\usage{
read_resource(package, resource_name, col_select = NULL)
}
\arguments{
\item{package}{Data Package object, as returned by \code{\link[frictionless:read_package]{read_package()}} or
\code{\link[frictionless:create_package]{create_package()}}.}

\item{resource_name}{Name of the Data Resource.}

\item{col_select}{Character vector of the columns to include in the result,
in the order provided.
Selecting columns can improve read speed.}
}
\description{
Read data from a Data Resource into a tibble data frame
}
\seealso{
Other frictionless functions: 
\code{\link{read_package}()},
\code{\link{write_package}()}
}
\concept{frictionless functions}
