% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_receiver_ids.R
\name{receiver_ids}
\alias{receiver_ids}
\title{List all available network receiver ids}
\usage{
receiver_ids(connection)
}
\arguments{
\item{connection}{A valid connection to the ETN database.}
}
\value{
A vector of all unique \code{receiver_id} present in \code{receivers}.
}
\description{
List all available network receiver ids
}
