% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frictionless-reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{read_package}
\alias{write_package}
\title{Objects exported from other packages}
\seealso{
Other frictionless functions: 
\code{\link{add_resource}()}

Other frictionless functions: 
\code{\link{add_resource}()}
}
\concept{frictionless functions}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{frictionless}{\code{\link[frictionless]{read_package}}, \code{\link[frictionless]{write_package}}}
}}

