% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{river_telemetry}
\alias{river_telemetry}
\title{Example ETN dataset with river telemetry data}
\format{
An object of class \code{datapackage} (inherits from \code{list}) of length 3.
}
\usage{
river_telemetry
}
\description{
An example ETN dataset "2014_DEMER", with river telemetry data, formatted as
a \href{https://specs.frictionlessdata.io/data-package/}{Frictionless Data Package} and read by
\code{\link[frictionless:read_package]{frictionless::read_package()}}.
}
\details{
\href{https://www.vliz.be/en/imis?dasid=5871&doiid=432}{"2014_DEMER"} is a dataset
of acoustic telemetry data collected for four fish species in the Demer
River, Belgium, in 2014. The example dataset is included in the ETN R
package, but can also be downloaded from the ETN database using the following
code:

\code{download_acoustic_dataset( animal_project_code = "2014_demer", directory = here::here("inst", "extdata") )}
}
\examples{
\dontrun{
# The data in river_telemetry was created with the code below
package_path <- system.file("extdata", "datapackage.json", package = "etn")
river_telemetry <- frictionless::read_package(package_path)
usethis::use_data(river_telemetry, overwrite = TRUE)
}
}
\concept{sample data}
\keyword{datasets}
