% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{scientific_names}
\alias{scientific_names}
\title{Get a list of unique available scientific_name}
\usage{
scientific_names(connection)
}
\arguments{
\item{connection}{A valid connection to the ETN database.}
}
\value{
A vector of all scientific_name present in vliz.animals_view2.
}
\description{
Get a list of unique available scientific_name
}
