% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_detections.R
\name{station_names}
\alias{station_names}
\title{Support function to get unique set of station_names}
\usage{
station_names(connection)
}
\arguments{
\item{connection}{A valid connection to the ETN database.}
}
\value{
A vector of all station_names present in vliz.deployments_view2.
}
\description{
Get unique station_names
}
