% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_detections.R
\name{tag_ids}
\alias{tag_ids}
\title{Support function to get unique set of tag_ids}
\usage{
tag_ids(connection)
}
\arguments{
\item{connection}{A valid connection to the ETN database.}
}
\value{
A vector of all tags present in vliz.tags_view2.
}
\description{
Get unique tag_ids
}
