% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_detections.R
\name{transmitters}
\alias{transmitters}
\title{Support function to get unique set of transmitters (tag_code_space)}
\usage{
transmitters(connection)
}
\arguments{
\item{connection}{A valid connection to ETN database.}
}
\value{
A vector of all transmitter present in vliz.tags.
}
\description{
Get unique transmitters
}
