% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{animal_projects}
\alias{animal_projects}
\title{Get a list of unique available animal_project_code}
\usage{
animal_projects(connection)
}
\arguments{
\item{connection}{A valid connection to the ETN database.}
}
\value{
A vector of all animal_project_code present in vliz.projects.
}
\description{
Get a list of unique available animal_project_code
}
