% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_connection}
\alias{check_connection}
\title{Support function to check the validity of the database connection}
\usage{
check_connection(connection)
}
\arguments{
\item{connection}{A valid connection with the ETN database.}
}
\description{
Support function to check the validity of the database connection
}
