% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_null_or_value}
\alias{check_null_or_value}
\title{Valid input is either NULL or option of list}
\usage{
check_null_or_value(arg, options = NULL, arg_name)
}
\arguments{
\item{arg}{The input argument provided by the user.}

\item{options}{A vector of valid inputs for the argument.}

\item{arg_name}{The name of the argument used in the function to test.}
}
\value{
If no error, TRUE.
}
\description{
Valid input is either NULL or option of list
}
\examples{
\dontrun{
# Invalid inputs for project_type
check_null_or_value("ddsf", c("animal", "network"), "project_type")
check_null_or_value("ddsf", c("animal", "network", "sea"), "project_type")

#Valid inputs for project_type
check_null_or_value("animal", c("animal", "network"), "project_type")
check_null_or_value(NULL, c("animal", "network"), "project_type")
check_null_or_value(c("animal", "network"), c("animal", "network"), "project_type")


# check network projects
valid_network_projects <- get_projects(connection = con,
                                       project_type = "network") \%>\%
                                       pull("projectcode")
check_null_or_value(c("thornton", "leopold"),
                    valid_network_projects, "network_project")
# check animal projects
valid_animal_projects <- get_projects(connection = con,
                                      project_type = "animal") \%>\%
                                      pull("projectcode")
check_null_or_value(c("2012_leopoldkanaal", "phd_reubens"),
                    valid_animal_projects, "animal_project")
}
}
