% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_null_or_value}
\alias{check_null_or_value}
\title{Valid input is either NULL or option of list}
\usage{
check_null_or_value(arg, options = NULL, arg_name)
}

\arguments{
\item{arg}{The input argument provided by the user.}

\item{options}{A vector of valid inputs for the argument.}

\item{arg_name}{The name of the argument used in the function to test.}
}
\value{
If no error, TRUE.
}

\examples{
check_null_or_value("ddsf", c("animal", "network"), "project_type")
check_null_or_value("ddsf", c("animal", "network", "sea"), "project_type")
check_null_or_value("animal", c("animal", "network"), "project_type")
check_null_or_value(NULL, c("animal", "network"), "project_type")
check_null_or_value(c("animal", "network"), c("animal", "network"), "project_type")
}
