% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_value}
\alias{check_value}
\title{Check input value against list of provided values}
\usage{
check_value(arg, options = NULL, arg_name)
}
\arguments{
\item{arg}{The input argument provided by the user.}

\item{options}{A vector of valid inputs for the argument.}

\item{arg_name}{The name of the argument used in the function to test.}
}
\value{
If no error, TRUE.
}
\description{
Will return error message if an input value cannot be found in list of provided
values. NULL values are allowed.
}
\examples{
\dontrun{
# Valid inputs for project_type
check_value("animal", c("animal", "network"), "project_type")
check_value(NULL, c("animal", "network"), "project_type")
check_value(c("animal", "network"), c("animal", "network"), "project_type")

# Invalid inputs for project_type
check_value("ddsf", c("animal", "network"), "project_type")
}
}
\keyword{internal}
