% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{collapse_transformer}
\alias{collapse_transformer}
\title{Print list of options}
\usage{
collapse_transformer(regex = "[*]$", ...)
}
\arguments{
\item{regex}{Character. A regular expression to parse.}

\item{...}{Additional arguments passed to the collapse.}
}
\description{
Print list of options
}
\keyword{internal}
