% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_etn.R
\name{connect_to_etn}
\alias{connect_to_etn}
\title{Connect to the ETN database}
\usage{
connect_to_etn(username, password)
}
\arguments{
\item{username}{Character. Username to use for the connection. Required.}

\item{password}{Character. Password to use for the connection. Required.}
}
\value{
ODBC connection to ETN database.
}
\description{
Connect to the ETN database using username and password.
}
\examples{
\dontrun{
con <- connect_to_etn(Sys.getenv("userid"), Sys.getenv("pwd"))
}
}
