% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_etn.R
\name{connect_to_etn}
\alias{connect_to_etn}
\title{Connect to the ETN database}
\usage{
connect_to_etn(...)
}
\arguments{
\item{...}{Any arguments passed to this function are ignored.}
}
\value{
This function is no longer in use, and returns NULL invisibly.
}
\description{
This function is \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} since etn version 2.3.0.
Its use is no longer supported or needed. All connections to the ETN database
are now made automatically when you use a function. If your credentials are not saved
in the system environement, you will be prompted to enter them.
}
