% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_etn.R
\name{connect_to_etn}
\alias{connect_to_etn}
\title{Connect to the database using username and password.}
\usage{
connect_to_etn(username, password)
}
\arguments{
\item{username}{(character) Username to use for the connection.}

\item{password}{(character) Password to use for the connection.}
}
\value{
con ODBC connection to ETN database.
}
\description{
Connect to the database using username and password.
}
\examples{
\dontrun{
con <- connect_to_etn(Sys.getenv("userid"), Sys.getenv("pwd"))
}
}
