% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_etn.R
\name{connect_to_etn}
\alias{connect_to_etn}
\title{Provide connection to the database with the user credentials}
\usage{
connect_to_etn(username, password)
}
\arguments{
\item{username}{char username to use the connection}

\item{password}{char password to use the connection}
}
\value{
conn ODBC connection to the database
}
\description{
Provide connection to the database with the user credentials
}
