% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_etn.R
\name{connect_to_etn}
\alias{connect_to_etn}
\title{Connect to the ETN database}
\usage{
connect_to_etn(username = Sys.getenv("userid"), password = Sys.getenv("pwd"))
}
\arguments{
\item{username}{Character. Username to use for the connection.}

\item{password}{Character. Password to use for the connection.}
}
\value{
ODBC connection to ETN database.
}
\description{
Connect to the ETN database using username and password.
}
\examples{
\dontrun{
# Connect to the ETN database using your rstudio.lifewatch.be username and
# password, and save as the default connection variable "con"
con <- connect_to_etn()

# Connect to the ETN database using non-default username and password
con <- connect_to_etn(username = "my_username", password = "my_password")
}
}
