% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\title{Download data package}
\usage{
download_dataset(
  connection = con,
  animal_project_code,
  directory = animal_project_code,
  scientific_name = NULL
)
}
\arguments{
\item{connection}{A valid connection to the ETN database. Defaults to \code{con}.}

\item{animal_project_code}{String. Animal project you want to download data
for. Required.}

\item{directory}{String. Relative path to local download directory.
Defaults to creating a directory named after animal project code.}

\item{scientific_name}{String. One or more scientific names to filter upon.
Defaults to no filter.}
}
\description{
Download all data related to an \strong{animal project} as a data package that can
be deposited in a research data repository. Includes option to filter on
scientific names.
}
\details{
The data are downloaded as a
\strong{\href{https://frictionlessdata.io/data-package/}{Frictionless Data Package}}
containing:\tabular{ll}{
   file \tab description \cr
   \code{animals.csv} \tab Animals related to an \code{animal_project_code}, optionally filtered on \code{scientific_name}(s), as returned by \code{get_animals()}. \cr
   \code{tags.csv} \tab Tags associated with the selected animals, as returned by \code{get_tags()}. \cr
   \code{detections.csv} \tab Detections for the selected animals, as returned by \code{get_detections()}. \cr
   \code{deployments.csv} \tab Deployments for the \code{network_project_code}(s) found in detections, as returned by \code{get_deployments()}. This allows users to see when receivers were deployed, even if these did not detect the selected animals. \cr
   \code{receivers.csv} \tab Receivers for the selected deployments, as returned by \code{get_receivers()}. \cr
   \code{datapackage.json} \tab A \href{https://specs.frictionlessdata.io/table-schema/}{Frictionless Table Schema} metadata file describing the fields and relations of the above csv files. This file is copied from \href{https://github.com/inbo/etn/blob/master/inst/assets/datapackage.json}{here} and can be used to validate the data package. \cr
}


The function will report the number of records per csv file, as well as the
included scientific names and network projects. Warnings will be raised for:
\itemize{
\item Animals with multiple tags
\item Tags associated with multiple animals
\item Deployments without network project: these deployments will not be listed
in \code{deployments.csv} and will therefore raise a foreign key validation error.
}

\strong{Important}: The data are downloaded \emph{as is} from the database, i.e. no
quality or consistency checks are performed by this function. We therefore
recommend to verify the data before publication. A consistency check can be
performed by validation tools of the Frictionless Framework, e.g.
\verb{frictionless validate datapackage.json} on the command line using
\href{https://github.com/frictionlessdata/frictionless-py}{frictionless-py}.
}
\examples{
\dontrun{
# Download data for the 2014_demer animal project (all scientific names)
download_dataset(con, animal_project_code = "2014_demer")
#> Downloading data to directory "2014_demer":
#> * (1/6): downloading animals.csv
#> * (2/6): downloading tags.csv
#> * (3/6): downloading detections.csv
#> * (4/6): downloading deployments.csv
#> * (5/6): downloading receivers.csv
#> * (6/6): adding datapackage.json as file metadata
#>
#> Summary statistics for dataset "2014_demer":
#> * number of animals:           16
#> * number of tags:              16
#> * number of detections:        237064
#> * number of deployments:       939
#> * number of receivers:         179
#> * included scientific names:   Petromyzon marinus, Rutilus rutilus, Silurus glanis, Squalius cephalus
#> * included network projects:   albert, demer, dijle, no_info, zeeschelde
}
}
