% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etn-deprecated.R
\name{etn-deprecated}
\alias{etn-deprecated}
\alias{get_deployments}
\alias{get_detections}
\alias{get_projects}
\alias{get_receivers}
\alias{list_network_project_codes}
\alias{list_tag_ids}
\title{Deprecated functions in etn}
\usage{
get_deployments(connection = con, network_project_code = NULL, ...)

get_detections(
  connection = con,
  tag_id = NULL,
  network_project_code = NULL,
  ...
)

get_projects(connection = con, project_type, application_type)

get_receivers(...)

list_network_project_codes(...)

list_tag_ids(...)
}
\description{
The functions listed below are deprecated or renamed and will be defunct in
the near future.
}
\keyword{internal}
