% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etn.R
\docType{package}
\name{etn-package}
\alias{etn}
\alias{etn-package}
\title{etn: Access Data from the European Tracking Network}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Package with functions to access and process data from the European
    Tracking Network hosted by VLIZ.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/inbo/etn}
  \item \url{https://inbo.github.io/etn}
  \item Report bugs at \url{https://github.com/inbo/etn/issues}
}

}
\author{
\strong{Maintainer}: Peter Desmet \email{peter.desmet@inbo.be} (\href{https://orcid.org/0000-0002-8442-8025}{ORCID})

Authors:
\itemize{
  \item Damiano Oldoni \email{damiano.oldoni@inbo.be} (\href{https://orcid.org/0000-0003-3445-7562}{ORCID})
  \item Stijn Van Hoey \email{stijnvanhoey@gmail.com} (\href{https://orcid.org/0000-0001-6413-3185}{ORCID})
}

Other contributors:
\itemize{
  \item Research Institute for Nature and Forest (INBO) (https://www.vlaanderen.be/inbo/en-gb/) [copyright holder]
  \item LifeWatch Belgium (https://lifewatch.be) [funder]
}

}
\keyword{internal}
