% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\name{example_dataset}
\alias{example_dataset}
\title{Read an example dataset}
\usage{
example_dataset(dataset = "2014_DEMER")
}
\arguments{
\item{dataset}{Name of the example dataset to load. Defaults to the only
available dataset: \code{"2014_DEMER"}.}
}
\value{
Frictionless data package.
}
\description{
Reads an example dataset, formatted as a \href{https://specs.frictionlessdata.io/data-package/}{Frictionless Data Package}.
}
\section{2014_DEMER}{


\verb{2014_DEMER} is a \strong{river acoustic telemetry} dataset. It contains 66 tagged
animals across four species. Over 235,000 detections were observed between
2014 and 2019 by acoustic receivers deployed in Belgian rivers. Data are
deposited at \url{https://doi.org/10.14284/432}.

The dataset was included in the package with:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{get_package(animal_project_code = "2014_demer") |>
  write_package("inst/extdata/2014_demer")
}\if{html}{\out{</div>}}
}

\examples{
example_dataset()
}
\concept{sample data}
