% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\name{example_dataset}
\alias{example_dataset}
\title{Read an example dataset}
\usage{
example_dataset(dataset = "2014_DEMER")
}
\arguments{
\item{dataset}{Name of the example dataset to load. Defaults to the only
available dataset: \code{"2014_DEMER"}.}
}
\value{
Frictionless data package.
}
\description{
An example ETN dataset, formatted as a \href{https://specs.frictionlessdata.io/data-package/}{Frictionless Data Package} and read by
\code{\link[frictionless:read_package]{frictionless::read_package()}}.
}
\details{
\href{https://www.vliz.be/en/imis?dasid=5871&doiid=432}{"2014_DEMER"} is a dataset
of acoustic river telemetry data collected for four fish species in the Demer
River, Belgium, in 2014. The example dataset is included in the ETN R
package, but can also be downloaded from the ETN database using the following
code:

\verb{download_acoustic_dataset( animal_project_code = "2014_DEMER", directory = here::here("inst", "extdata", "2014_DEMER")) )}
}

\examples{
example_dataset()
}
\concept{sample data}
