% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acoustic_deployment_logs.R
\name{get_acoustic_deployment_logs}
\alias{get_acoustic_deployment_logs}
\title{Get diagnostics information for a receiver for a deployment id.}
\usage{
get_acoustic_deployment_logs(deployment_id, limit = FALSE)
}
\arguments{
\item{deployment_id}{Integer (vector). One or more deployment identifiers.}

\item{limit}{Logical. Limit the number of returned records to 100 (useful
for testing purposes). Defaults to \code{FALSE}.}
}
\value{
A tibble with receiver diagnostics data.
}
\description{
This function retrieves diagnostics information for a receiver for a
deployment id. This may be useful to figure out what happened to a receiver
during deployment. For example, a deviation in tilt angle may decrease
detection capabilities. For some receivers, information about the water
temperature can be available.
}
\details{
The diagnostics information is returned as a tibble with one
row per log entry. The columns of the tibble may vary depending on the
deployment and receiver. If no log entries are found for the specified
arguments, an 0-row tibble is returned.
}
\section{Name repair}{


It is possible that the columns contained in a receiver log overlap with
the default columns always returned by \code{get_acoustic_deployment_logs()}. If
duplicate columns are found, their names are made unique with
\code{\link[base:make.unique]{base::make.unique()}}. When this happens, a message is printed to the
console. The message can be muffled with \code{\link[base:suppressMessages]{base::suppressMessages()}}.
}

\examples{
\dontshow{if (etn:::credentials_are_set()) withAutoprint(\{ # examplesIf}
get_acoustic_deployment_logs(deployment_id = 25259, limit = TRUE)
get_acoustic_deployment_logs(deployment_id = 74535,
                  station_name = "Tijdenskan. - Hamdijk 12 Bad Nieuwesch")
get_acoustic_deployment_logs(deployment_id = 6028,
                  start_date = "2020",
                  end_date = "2020-02-01")
\dontshow{\}) # examplesIf}
}
\seealso{
Other access functions:
\code{\link[=get_acoustic_deployments]{get_acoustic_deployments()}},
\code{\link[=get_acoustic_detections]{get_acoustic_detections()}},
\code{\link[=get_acoustic_projects]{get_acoustic_projects()}},
\code{\link[=get_acoustic_receivers]{get_acoustic_receivers()}},
\code{\link[=get_animal_projects]{get_animal_projects()}},
\code{\link[=get_animals]{get_animals()}},
\code{\link[=get_cpod_projects]{get_cpod_projects()}},
\code{\link[=get_package]{get_package()}},
\code{\link[=get_tags]{get_tags()}}
}
\concept{access functions}
