% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acoustic_deployments.R
\name{get_acoustic_deployments}
\alias{get_acoustic_deployments}
\title{Get acoustic deployment data}
\usage{
get_acoustic_deployments(
  connection = con,
  receiver_id = NULL,
  network_project_code = NULL,
  open_only = FALSE
)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}

\item{receiver_id}{Character (vector). One or more receiver identifiers.}

\item{network_project_code}{Character (vector). One or more network
projects. Case-insensitive.}

\item{open_only}{Logical. Restrict deployments to those that are currently
open (i.e. no end date defined). Defaults to \code{FALSE}.}
}
\value{
A tibble with acoustic deployment data, sorted by
\code{network_project_code}, \code{station_name} and \code{deploy_date_time}. See also
\href{https://inbo.github.io/etn/articles/etn_fields.html}{field definitions}.
}
\description{
Get data for deployments of acoustic receivers, with options to filter
results.
}
\examples{
\dontrun{
# Set default connection variable
con <- connect_to_etn()

# Get all acoustic deployments
get_acoustic_deployments()

# Get open acoustic deployments (i.e. no end date)
get_acoustic_deployments(open_only = TRUE)

# Get all deployments for a specific network project
get_acoustic_deployments(con, network_project_code = "demer")
}
}
