% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acoustic_detections.R
\name{get_acoustic_detections}
\alias{get_acoustic_detections}
\title{Get acoustic detections data}
\usage{
get_acoustic_detections(
  connection,
  start_date = NULL,
  end_date = NULL,
  tag_serial_number = NULL,
  acoustic_tag_id = NULL,
  animal_project_code = NULL,
  scientific_name = NULL,
  acoustic_project_code = NULL,
  receiver_id = NULL,
  station_name = NULL,
  limit = FALSE,
  progress = TRUE
)
}
\arguments{
\item{connection}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A connection to the ETN
database. This argument is no longer used. You will be prompted for
credentials instead.}

\item{start_date}{Character. Start date (inclusive) in ISO 8601 format (
\code{yyyy-mm-dd}, \code{yyyy-mm} or \code{yyyy}).}

\item{end_date}{Character. End date (exclusive) in ISO 8601 format (
\code{yyyy-mm-dd}, \code{yyyy-mm} or \code{yyyy}).}

\item{tag_serial_number}{Character (vector). One or more acoustic tag serial
numbers.}

\item{acoustic_tag_id}{Character (vector). One or more acoustic tag ids.}

\item{animal_project_code}{Character (vector). One or more animal project
codes. Case-insensitive.}

\item{scientific_name}{Character (vector). One or more scientific names.}

\item{acoustic_project_code}{Character (vector). One or more acoustic
project codes. Case-insensitive.}

\item{receiver_id}{Character (vector). One or more receiver identifiers.}

\item{station_name}{Character (vector). One or more deployment station
names.}

\item{limit}{Logical. Limit the number of returned records to 100 (useful
for testing purposes). Defaults to \code{FALSE}.}

\item{progress}{Logical. Show a progress bar while fetching data. Defaults to
\code{TRUE}.}
}
\value{
A tibble with acoustic detections data, sorted by \code{acoustic_tag_id}
and \code{date_time}.
}
\description{
Get data for acoustic detections, with options to filter results. Use
\code{limit} to limit the number of returned records.
}
\examples{
\dontshow{if (etn:::credentials_are_set()) withAutoprint(\{ # examplesIf}
# Get limited sample of acoustic detections
get_acoustic_detections(limit = TRUE)

# Get all acoustic detections from a specific animal project
get_acoustic_detections(animal_project_code = "2014_demer")

# Get 2015 acoustic detections from that animal project
get_acoustic_detections(
  animal_project_code = "2014_demer",
  start_date = "2015",
  end_date = "2016",
)

# Get April 2015 acoustic detections from that animal project
get_acoustic_detections(
  animal_project_code = "2014_demer",
  start_date = "2015-04",
  end_date = "2015-05",
)

# Get April 24, 2015 acoustic detections from that animal project
get_acoustic_detections(
  animal_project_code = "2014_demer",
  start_date = "2015-04-24",
  end_date = "2015-04-25",
)

# Get acoustic detections for a specific tag at two specific stations
get_acoustic_detections(
  acoustic_tag_id = "A69-1601-16130",
  station_name = c("de-9", "de-10")
)

# Get acoustic detections for a specific species, receiver and acoustic project
get_acoustic_detections(
  scientific_name = "Rutilus rutilus",
  receiver_id = "VR2W-124070",
  acoustic_project_code = "demer"
)
\dontshow{\}) # examplesIf}
}
\seealso{
Other access functions:
\code{\link[=get_acoustic_deployment_logs]{get_acoustic_deployment_logs()}},
\code{\link[=get_acoustic_deployments]{get_acoustic_deployments()}},
\code{\link[=get_acoustic_projects]{get_acoustic_projects()}},
\code{\link[=get_acoustic_receivers]{get_acoustic_receivers()}},
\code{\link[=get_animal_projects]{get_animal_projects()}},
\code{\link[=get_animals]{get_animals()}},
\code{\link[=get_bibliography]{get_bibliography()}},
\code{\link[=get_cpod_projects]{get_cpod_projects()}},
\code{\link[=get_package]{get_package()}},
\code{\link[=get_tags]{get_tags()}}
}
\concept{access functions}
