% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acoustic_projects.R
\name{get_acoustic_projects}
\alias{get_acoustic_projects}
\title{Get acoustic project data}
\usage{
get_acoustic_projects(connection, acoustic_project_code = NULL)
}
\arguments{
\item{connection}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A connection to the ETN
database. This argument is no longer used. You will be prompted for
credentials instead.}

\item{acoustic_project_code}{Character (vector). One or more acoustic
project codes. Case-insensitive.}
}
\value{
A tibble with acoustic project data, sorted by \code{project_code}. See
also
\href{https://inbo.github.io/etn/articles/etn_fields.html}{field definitions}.
}
\description{
Get data for acoustic projects, with options to filter results.
}
\examples{
# Get all acoustic projects
get_acoustic_projects()

# Get a specific acoustic project
get_acoustic_projects(acoustic_project_code = "demer")
}
