% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acoustic_receivers.R
\name{get_acoustic_receivers}
\alias{get_acoustic_receivers}
\title{Get acoustic receiver data}
\usage{
get_acoustic_receivers(connection, receiver_id = NULL, status = NULL)
}
\arguments{
\item{connection}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A connection to the ETN
database. This argument is no longer used. You will be prompted for
credentials instead.}

\item{receiver_id}{Character (vector). One or more receiver identifiers.}

\item{status}{Character. One or more statuses, e.g. \code{available} or \code{broken}.}
}
\value{
A tibble with acoustic receiver data, sorted by \code{receiver_id}.
Values for \code{owner_organization} will only be visible if you are member of
the group.
}
\description{
Get data for acoustic receivers, with options to filter results.
}
\examples{
\dontshow{if (etn:::credentials_are_set()) withAutoprint(\{ # examplesIf}
# Get all acoustic receivers
get_acoustic_receivers()

# Get lost and broken acoustic receivers
get_acoustic_receivers(status = c("lost", "broken"))

# Get a specific acoustic receiver
get_acoustic_receivers(receiver_id = "VR2W-124070")
\dontshow{\}) # examplesIf}
}
