% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acoustic_receivers.R
\name{get_acoustic_receivers}
\alias{get_acoustic_receivers}
\title{Get acoustic receiver data}
\usage{
get_acoustic_receivers(connection = con, receiver_id = NULL, status = NULL)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}

\item{receiver_id}{Character (vector). One or more receiver identifiers.}

\item{status}{Character. One or more statuses, e.g. \code{available} or \code{broken}.}
}
\value{
A tibble with acoustic receiver data, sorted by \code{receiver_id}. See
also
\href{https://inbo.github.io/etn/articles/etn_fields.html}{field definitions}.
Values for \code{owner_organization} will only be visible if you are member of
the group.
}
\description{
Get data for acoustic receivers, with options to filter results.
}
\examples{
# Set default connection variable
con <- connect_to_etn()

# Get all acoustic receivers
get_acoustic_receivers(con)

# Get lost and broken acoustic receivers
get_acoustic_receivers(con, status = c("lost", "broken"))

# Get a specific acoustic receiver
get_acoustic_receivers(con, receiver_id = "VR2W-124070")
}
