% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acoustic_tags.R
\name{get_acoustic_tags}
\alias{get_acoustic_tags}
\title{Get acoustic tag data}
\usage{
get_acoustic_tags(
  connection = con,
  tag_serial_number = NULL,
  acoustic_tag_id = NULL
)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}

\item{tag_serial_number}{Character (vector). One or more tag serial numbers.}

\item{acoustic_tag_id}{Character (vector). One or more acoustic tag
identifiers.}
}
\value{
A tibble with tags data, sorted by \code{tag_serial_number}. See also
\href{https://inbo.github.io/etn/articles/etn_fields.html}{field definitions}.
}
\description{
Get data for acoustic tags, with options to filter results. Note that there
can be multiple records (\code{acoustic_tag_id}) per tag device
(\code{tag_serial_number}).
}
\examples{
\dontrun{
# Set default connection variable
con <- connect_to_etn()

# Get all acoustic tags
get_acoustic_tags()

# Get specific acoustic tags
get_acoustic_tags(tag_serial_number = "1187450")
get_acoustic_tags(acoustic_tag_id = "A69-1601-16130")
get_acoustic_tags(acoustic_tag_id = c("A69-1601-16129", "A69-1601-16130"))
}
}
