% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_animal_projects.R
\name{get_animal_projects}
\alias{get_animal_projects}
\title{Get animal project data}
\usage{
get_animal_projects(connection, animal_project_code = NULL, citation = FALSE)
}
\arguments{
\item{connection}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A connection to the ETN
database. This argument is no longer used. You will be prompted for
credentials instead.}

\item{animal_project_code}{Character (vector). One or more animal project
codes. Case-insensitive.}

\item{citation}{Logical. If \code{TRUE}, adds extra columns with citation
information for each project from \href{https://marineinfo.org/}{MarineInfo}
using the \code{imis_dataset_id}:
\itemize{
\item \code{citation}: Formatted citation with DOI if available.
\item \code{doi}: DOI for the dataset if available.
\item \code{contact_name}: Contact person, usually the first author.
If no contact person is provided, the first author with status \code{creator}.
\item \code{contact_email}: Email address of the contact person.
\item \code{contact_affiliation}: Institute of the contact person.
}}
}
\value{
A tibble with animal project data, sorted by \code{project_code}.
}
\description{
Get data for animal projects, with options to filter results.
}
\examples{
\dontshow{if (etn:::credentials_are_set()) withAutoprint(\{ # examplesIf}
# Get all animal projects
get_animal_projects()

# Get a specific animal project with citation
get_animal_projects(animal_project_code = "2014_demer", citation = TRUE)
\dontshow{\}) # examplesIf}
}
\seealso{
Other access functions:
\code{\link[=get_acoustic_deployment_logs]{get_acoustic_deployment_logs()}},
\code{\link[=get_acoustic_deployments]{get_acoustic_deployments()}},
\code{\link[=get_acoustic_detections]{get_acoustic_detections()}},
\code{\link[=get_acoustic_projects]{get_acoustic_projects()}},
\code{\link[=get_acoustic_receivers]{get_acoustic_receivers()}},
\code{\link[=get_animals]{get_animals()}},
\code{\link[=get_cpod_projects]{get_cpod_projects()}},
\code{\link[=get_package]{get_package()}},
\code{\link[=get_tags]{get_tags()}}
}
\concept{access functions}
