% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_animal_projects.R
\name{get_animal_projects}
\alias{get_animal_projects}
\title{Get animal project data}
\usage{
get_animal_projects(connection = con, animal_project_code = NULL)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}

\item{animal_project_code}{Character (vector). One or more animal project
codes. Case-insensitive.}
}
\value{
A tibble with animal project data, sorted by \code{project_code}. See
also
\href{https://inbo.github.io/etn/articles/etn_fields.html}{field definitions}.
}
\description{
Get data for animal projects, with options to filter results.
}
\examples{
# Set default connection variable
con <- connect_to_etn()

# Get all animal projects
get_animal_projects(con)

# Get a specific animal project
get_animal_projects(con, animal_project_code = "2014_demer")
}
