% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_animals.R
\name{get_animals}
\alias{get_animals}
\title{Get animal data}
\usage{
get_animals(
  connection,
  animal_id = NULL,
  tag_serial_number = NULL,
  animal_project_code = NULL,
  scientific_name = NULL
)
}
\arguments{
\item{connection}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A connection to the ETN
database. This argument is no longer used. You will be prompted for
credentials instead.}

\item{animal_id}{Integer (vector). One or more animal identifiers.}

\item{tag_serial_number}{Character (vector). One or more tag serial numbers.}

\item{animal_project_code}{Character (vector). One or more animal project
codes. Case-insensitive.}

\item{scientific_name}{Character (vector). One or more scientific names.}
}
\value{
A tibble with animals data, sorted by \code{animal_project_code},
\code{release_date_time} and \code{tag_serial_number}.
}
\description{
Get data for animals, with options to filter results. Associated tag
information is available in columns starting with \code{tag} and
\code{acoustic_tag_id}. If multiple tags are associated with a single animal,
the information is comma-separated.
}
\examples{
\dontshow{if (etn:::credentials_are_set()) withAutoprint(\{ # examplesIf}
# Get all animals
get_animals()

# Get specific animals
get_animals(animal_id = 305) # Or string value "305"
get_animals(animal_id = c(304, 305, 2827))

# Get animals from specific animal project(s)
get_animals(animal_project_code = "2014_demer")
get_animals(animal_project_code = c("2014_demer", "2015_dijle"))

# Get animals associated with a specific tag_serial_number
get_animals(tag_serial_number = "1187450")

# Get animals of specific species (across all projects)
get_animals(scientific_name = c("Rutilus rutilus", "Silurus glanis"))

# Get animals of a specific species from a specific project
get_animals(animal_project_code = "2014_demer", scientific_name = "Rutilus rutilus")
\dontshow{\}) # examplesIf}
}
\seealso{
Other access functions:
\code{\link[=get_acoustic_deployment_logs]{get_acoustic_deployment_logs()}},
\code{\link[=get_acoustic_deployments]{get_acoustic_deployments()}},
\code{\link[=get_acoustic_detections]{get_acoustic_detections()}},
\code{\link[=get_acoustic_projects]{get_acoustic_projects()}},
\code{\link[=get_acoustic_receivers]{get_acoustic_receivers()}},
\code{\link[=get_animal_projects]{get_animal_projects()}},
\code{\link[=get_bibliography]{get_bibliography()}},
\code{\link[=get_cpod_projects]{get_cpod_projects()}},
\code{\link[=get_package]{get_package()}},
\code{\link[=get_tags]{get_tags()}}
}
\concept{access functions}
