% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_animals.R
\name{get_animals}
\alias{get_animals}
\title{Get animals data}
\usage{
get_animals(connection, network_project = NULL, animal_project = NULL,
  scientific_name = NULL)
}
\arguments{
\item{connection}{A valid connection to ETN database.}

\item{network_project}{(string) One or more network projects.}

\item{animal_project}{(string) One or more animal projects.}

\item{scientific_name}{(string) One or more scientific names.}
}
\value{
A data.frame.
}
\description{
This function retrieves all or a subset of animals data.
}
\examples{
\dontrun{
# all animals
animals <- get_animals(con)

# all animals of the Demer project with given scientific name
animals <- get_animals(con, network_project = "demer",
                       scientific_name = c("Rutilus rutilus",
                       "Squalius cephalus"))
}
}
