% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_animals.R
\name{get_animals}
\alias{get_animals}
\title{Get animal metadata}
\usage{
get_animals(connection, animal_project = NULL, scientific_name = NULL)
}
\arguments{
\item{connection}{A valid connection to ETN database.}

\item{animal_project}{(string) One or more animal projects.}

\item{scientific_name}{(string) One or more scientific names.}
}
\value{
A tibble (tidyverse data.frame).
}
\description{
Get metadata for animals, with options to filter on animal project and/or
scientific name.
}
\examples{
\dontrun{
con <- connect_to_etn(your_username, your_password)

# Get metadata for all animals
animals <- get_animals(con)

# Get metadata for animals linked to specific project(s)
animals <- get_animals(con, animal_project = "2012_leopoldkanaal")
animals <- get_animals(con, animal_project = c("2012_leopoldkanaal", "phd_reubens"))

# Get metadata for animals with specific scientific names (across all projects)
animals <- get_animals(con, scientific_name = c("Gadus morhua", "Sentinel"))

# Get metadata for animals with specific scientific name for specific project
animals <- get_animals(con, animal_project = "phd_reubens", scientific_name = "Gadus morhua")
}
}
