% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_animals.R
\name{get_animals}
\alias{get_animals}
\title{Get animal data}
\usage{
get_animals(
  animal_id = NULL,
  tag_serial_number = NULL,
  animal_project_code = NULL,
  scientific_name = NULL,
  api = TRUE,
  connection
)
}
\arguments{
\item{animal_id}{Integer (vector). One or more animal identifiers.}

\item{tag_serial_number}{Character (vector). One or more tag serial numbers.}

\item{animal_project_code}{Character (vector). One or more animal project
codes. Case-insensitive.}

\item{scientific_name}{Character (vector). One or more scientific names.}

\item{api}{Logical. If \code{TRUE}, the function will use the API to access ETN.}

\item{connection}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A connection to the ETN
database. This argument is no longer used. You will be prompted for
credentials instead.}
}
\value{
A tibble with animals data, sorted by \code{animal_project_code},
\code{release_date_time} and \code{tag_serial_number}. See also
\href{https://inbo.github.io/etn/articles/etn_fields.html}{field definitions}.
}
\description{
Get data for animals, with options to filter results. Associated tag
information is available in columns starting with \code{tag} and
\code{acoustic_tag_id}. If multiple tags are associated with a single animal,
the information is comma-separated.
}
\examples{
# Set default connection variable
con <- connect_to_etn()

# Get all animals
get_animals(con)

# Get specific animals
get_animals(con, animal_id = 305) # Or string value "305"
get_animals(con, animal_id = c(304, 305, 2827))

# Get animals from specific animal project(s)
get_animals(con, animal_project_code = "2014_demer")
get_animals(con, animal_project_code = c("2014_demer", "2015_dijle"))

# Get animals associated with a specific tag_serial_number
get_animals(con, tag_serial_number = "1187450")

# Get animals of specific species (across all projects)
get_animals(con, scientific_name = c("Rutilus rutilus", "Silurus glanis"))

# Get animals of a specific species from a specific project
get_animals(con, animal_project_code = "2014_demer", scientific_name = "Rutilus rutilus")
}
