% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_animals.R
\name{get_animals}
\alias{get_animals}
\title{Get animal metadata}
\usage{
get_animals(
  connection = con,
  animal_id = NULL,
  animal_project_code = NULL,
  scientific_name = NULL
)
}
\arguments{
\item{connection}{A valid connection to the ETN database.}

\item{animal_id}{(integer) One or more animal ids.}

\item{animal_project_code}{(string) One or more animal projects.}

\item{scientific_name}{(string) One or more scientific names.}
}
\value{
A tibble (tidyverse data.frame).
}
\description{
Get metadata for animals, with options to filter on animal project and/or
scientific name. Associated tag information is available in columns starting
with \code{tag}. If multiple tags are associated with a single animal, the
information is comma-separated.
}
\examples{
\dontrun{
con <- connect_to_etn(your_username, your_password)

# Get all animals
get_animals(con)

# Get specific animals
get_animals(con, animal_id = 304)
get_animals(con, animal_id = c(304, 305))

# Get animals from specific animal project(s)
get_animals(con, animal_project_code = "2012_leopoldkanaal")
get_animals(con, animal_project_code = c("2012_leopoldkanaal", "phd_reubens"))

# Get animals of specific species (across all projects)
get_animals(con, scientific_name = c("Gadus morhua", "Sentinel"))

# Get animals of a specific species from a specific project
get_animals(con, animal_project_code = "phd_reubens", scientific_name = "Gadus morhua")
}
}
