% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_animals.R
\name{get_animals}
\alias{get_animals}
\title{Get animal data}
\usage{
get_animals(
  connection = con,
  animal_id = NULL,
  animal_project_code = NULL,
  scientific_name = NULL
)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}

\item{animal_id}{Integer (vector). One or more animal ids.}

\item{animal_project_code}{Character (vector). One or more animal projects.}

\item{scientific_name}{Character (vector). One or more scientific names.}
}
\value{
A tibble with animals data, sorted by \code{animal_project_code},
\code{release_date_time} and \code{tag_id}. See also
\href{https://inbo.github.io/etn/articles/etn_fields.html}{field definitions}.
}
\description{
Get data for animals, with options to filter results. Associated tag
information is available in columns starting with \code{tag}. If multiple tags
are associated with a single animal, the information is comma-separated.
}
\examples{
\dontrun{
con <- connect_to_etn(your_username, your_password)

# Get all animals
get_animals(con)

# Get specific animals
get_animals(con, animal_id = 2824)
get_animals(con, animal_id = "2824") # String values work as well
get_animals(con, animal_id = c(2824, 2825, 2827))

# Get animals from specific animal project(s)
get_animals(con, animal_project_code = "2012_leopoldkanaal")
get_animals(con, animal_project_code = c("2012_leopoldkanaal", "phd_reubens"))

# Get animals of specific species (across all projects)
get_animals(con, scientific_name = c("Gadus morhua", "Sentinel"))

# Get animals of a specific species from a specific project
get_animals(con, animal_project_code = "phd_reubens", scientific_name = "Gadus morhua")
}
}
