% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_animals.R
\name{get_animals}
\alias{get_animals}
\title{Get animal data}
\usage{
get_animals(
  connection = con,
  animal_id = NULL,
  animal_project_code = NULL,
  scientific_name = NULL,
  exclude_non_animals = FALSE
)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}

\item{animal_id}{Integer (vector). One or more animal ids.}

\item{animal_project_code}{Character (vector). One or more animal projects.
Case-insensitive.}

\item{scientific_name}{Character (vector). One or more scientific names.}

\item{exclude_non_animals}{Logical, \code{FALSE} by default. Exclude records with
non-animal scientific names such as "Built-in" or "Sync tag".}
}
\value{
A tibble with animals data, sorted by \code{animal_project_code},
\code{release_date_time} and \code{tag_serial_number}. See also
\href{https://inbo.github.io/etn/articles/etn_fields.html}{field definitions}.
}
\description{
Get data for animals, with options to filter results. Associated tag
information is available in columns starting with \code{tag}. If multiple tags
are associated with a single animal, the information is comma-separated.
}
\examples{
\dontrun{
con <- connect_to_etn()

# Get all animals
get_animals(con)

# Get specific animals
get_animals(con, animal_id = 384) # Or string value "384"
get_animals(con, animal_id = c(384, 385, 2827))

# Get animals from specific animal project(s)
get_animals(con, animal_project_code = "2014_demer")
get_animals(con, animal_project_code = c("2014_demer", "2015_dijle"))

# Get animals of specific species (across all projects)
get_animals(con, scientific_name = c("Silurus glanis", "Petromyzon marinus"))

# Get animals of a specific species from a specific project
get_animals(con, animal_project_code = "2014_demer", scientific_name = "Silurus glanis")

# Exclude non-animals (e.g. Sync tag)
get_animals(exclude_non_animals = TRUE)
}
}
