% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_animals.R
\name{get_animals}
\alias{get_animals}
\title{Get animals data}
\usage{
get_animals(connection, network_project = NULL, animal_project = NULL,
  scientific_name = NULL)
}
\arguments{
\item{connection}{A valid connection to ETN database.}

\item{network_project}{(string) One or more network projects.}

\item{animal_project}{(string) One or more animal projects.}

\item{scientific_name}{(string) One or more scientific names.}
}
\value{
A data.frame.
}
\description{
Get all or a subset of animals data. Filter options are the network
project(s), the animal project(s) and/or the scientific name of the animals.
}
\examples{
\dontrun{
con <- connect_to_etn(your_username, your_password)

# all animals
animals <- get_animals(con)

# all animals of the 2012_leopoldkanaal project
animals <- get_animals(con, animal_project = "2012_leopoldkanaal")

# all animals of the 2012_leopoldkanaal and phd_reubens
get_animals(con, animal_project = c("2012_leopoldkanaal", "phd_reubens"))

#all animals of the thornton network project
get_animals(con, network_project = "thornton")

#all animals of the thornton and leopold network project
get_animals(con, network_project = c("thornton", "leopold"))

# all animals of all projects with given scientific name
animals <- get_animals(con,
                       scientific_name = c("Gadus morhua",
                                           "Sentinel"))

# all animals of the phd_reubens project with given scientific name
animals <- get_animals(con, animal_project = "phd_reubens",
                       scientific_name = "Gadus morhua")
}
}
