% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archival_data.R
\name{get_archival_data}
\alias{get_archival_data}
\title{Get processed tag archival data}
\usage{
get_archival_data(
  tag_serial_number = NULL,
  animal_id = NULL,
  animal_project_code = NULL,
  path = NULL,
  limit = FALSE,
  progress = TRUE,
  return_as = c("tibble", "arrow")
)
}
\arguments{
\item{tag_serial_number}{Character (vector). One or more tag serial numbers.}

\item{animal_id}{Integer (vector). One or more animal identifiers.}

\item{animal_project_code}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Character
(vector). One or more animal project codes. Case-insensitive. An animal
project can contain many gigabytes of archival data. Downloading this data
may take a while and  exceed the available RAM or Storage of your computer.}

\item{path}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Character. A path to an
existing directory where the csv files will be stored. If NULL, the csv
files will be stored in a temporary directory that is deleted when R is
restarted. If you want to keep the csv files, provide a path. Keep in mind
that the csv files can be quite large and may take up a lot of disk space.
If the requested files are already present in the provided path, they will
not be downloaded again.}

\item{limit}{Logical. Limit the number of returned records to 100 (useful
for testing purposes). Defaults to \code{FALSE}.}

\item{progress}{Logical. Show a progress bar while fetching data. Defaults to
\code{TRUE}.}

\item{return_as}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Character. One of
"tibble" or "arrow". Whether to return the data as an in memory tibble or
an out of memory arrow dataset. The latter is recommended for large
datasets, but requires more disk space and may be slower to query. When
selecting "arrow" keep in mind that the downloaded data will remain stored
on your computer until R is restarted.}
}
\value{
Depending on the value of \code{return_as}, either an \link[arrow:Dataset]{arrow::Dataset} or
a data.frame with the archival data values. The resulting table has the
following columns:
\itemize{
\item \code{tag_id}: The tag id of the tag that recorded the data.
\item \code{timestamp_utc}: The timestamp of the measurement in UTC.
\item \code{measurement_type}: The type of measurement (e.g. "Temp", "Pressure, etc.).
\item \code{measurement_value}: The value of the measurement.
\item \code{measurement_unit}: The unit of the measurement (e.g. "C", "Pa", etc.).
\item \code{animal_project_code}: The animal project code of the animal that recorded the data.
\item \code{animal_id}: The animal id of the animal that recorded the data.
\item \code{tag_serial_number}: The tag serial number of the tag that recorded the data.
}
}
\description{
This function fetches processed archival data from tags. The data is stored
in csv files hosted on lifewatch.com. Providing at least one of the following
arguments is required: \code{tag_serial_number}, \code{animal_id}, or
\code{animal_project_code}.
}
\section{Returning large amounts of data out of memory}{

If you are fetching a large amount of data, it is recommended to use
\code{return_as = "arrow"} to return an out of memory object. This will allow you
to query the data without loading it all into memory at once. Keep in mind
that the downloaded data will remain stored on your computer until R is
restarted. If you want to keep the data, you can provide a path to an
existing directory where the csv files will be stored via the \code{path}
argument. See \link[arrow:Dataset]{arrow::Dataset} for more information.
}

\examples{
\dontshow{if (interactive() & etn:::credentials_are_set()) withAutoprint(\{ # examplesIf}
  get_archival_data(tag_serial_number = "A15757", limit = TRUE)
\dontshow{\}) # examplesIf}
}
\seealso{
Other access functions:
\code{\link[=get_acoustic_deployments]{get_acoustic_deployments()}},
\code{\link[=get_acoustic_detections]{get_acoustic_detections()}},
\code{\link[=get_acoustic_projects]{get_acoustic_projects()}},
\code{\link[=get_acoustic_receivers]{get_acoustic_receivers()}},
\code{\link[=get_animal_projects]{get_animal_projects()}},
\code{\link[=get_animals]{get_animals()}},
\code{\link[=get_cpod_projects]{get_cpod_projects()}},
\code{\link[=get_package]{get_package()}},
\code{\link[=get_tags]{get_tags()}}
}
\concept{access functions}
