% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archival_data.R
\name{get_archival_data}
\alias{get_archival_data}
\title{Get processed tag archival data}
\usage{
get_archival_data(
  tag_serial_number = NULL,
  animal_id = NULL,
  animal_project_code = NULL,
  path = NULL,
  limit = FALSE,
  progress = TRUE,
  return_as = c("tibble", "arrow")
)
}
\arguments{
\item{tag_serial_number}{Character (vector). One or more tag serial numbers.}

\item{animal_id}{Integer (vector). One or more animal identifiers.}

\item{animal_project_code}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Character
(vector). One or more animal project codes. Case-insensitive. An animal
project can contain many gigabytes of archival data. Downloading this data
may take a while and  exceed the available RAM or Storage of your computer.}

\item{path}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Character. A path to a
directory where the csv files will be stored. If NULL, the csv files will
be stored in a temporary directory that is deleted when R is restarted. If
you want to keep the csv files, provide a path. Keep in mind that the csv
files can be quite large and may take up a lot of disk space.}

\item{limit}{Logical. Limit the number of returned records to 100 (useful
for testing purposes). Defaults to \code{FALSE}.}

\item{progress}{Logical. Show a progress bar while fetching data. Defaults to
\code{TRUE}.}

\item{return_as}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Character. One of
"tibble" or "arrow". Whether to return the data as an in memory tibble or
an out of memory arrow dataset. The latter is recommended for large
datasets, but requires more disk space and may be slower to query. When
selecting "arrow" keep in mind that the downloaded data will remain stored
on your computer until R is restarted.}
}
\value{
Depending on the value of \code{return_as}, either an \link[arrow:Dataset]{arrow::Dataset} or a data.frame with the archival data
values. This data.frame has the following columns:
\itemize{
\item \code{tag_id}: The tag id of the tag that recorded the data.
\item \code{timestamp_utc}: The timestamp of the measurement in UTC.
\item \code{measurement_type}: The type of measurement (e.g. "temperature", "depth", etc.).
\item \code{measurement_value}: The value of the measurement.
\item \code{measurement_unit}: The unit of the measurement (e.g. "Pa", "m", etc.).
\item \code{animal_project_code}: The animal project code of the animal that recorded the data.
\item \code{animal_id}: The animal id of the animal that recorded the data.
\item \code{tag_serial_number}: The tag serial number of the tag that recorded the data.
When \code{path} is set to a value other than NULL. The returned data.frame will
not contain the \code{animal_project_code}, \code{animal_id}, and \code{tag_serial_number}
columns, since these are only available in the metadata that is used to
fetch the data, and not in the csv files that are downloaded. When
\code{return_as} is set to "arrow", an arrow dataset is returned instead of a
tibble. This dataset can be queried using dplyr functions, but cannot be
printed in its entirety.
}
}
\description{
Get processed tag archival data
}
\examples{
\dontshow{if (interactive() & credentials_are_set()) withAutoprint(\{ # examplesIf}
  get_archival_data(tag_serial_number = "A15757", limit = TRUE)
\dontshow{\}) # examplesIf}
}
