% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archival_data.R
\name{get_archival_data}
\alias{get_archival_data}
\title{Get processed tag archival data}
\usage{
get_archival_data(
  tag_serial_number = NULL,
  animal_id = NULL,
  animal_project_code = NULL
)
}
\arguments{
\item{tag_serial_number}{Character (vector). One or more tag serial numbers.}

\item{animal_id}{Integer (vector). One or more animal identifiers.}

\item{animal_project_code}{Character (vector). One or more animal project
codes. Case-insensitive.}
}
\value{
A data.frame with the archival data values.
}
\description{
Get processed tag archival data
}
\examples{
\dontshow{if (interactive() & credentials_are_set()) withAutoprint(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
