% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bibliography.R
\name{get_bibliography}
\alias{get_bibliography}
\title{Create bibliography from detections}
\usage{
get_bibliography(detections)
}
\arguments{
\item{detections}{A data frame containing at least the columns \code{animal_project_code}
and \code{acoustic_project_code}. Typically a data frame returned by
\code{\link[=get_acoustic_detections]{get_acoustic_detections()}}.}
}
\value{
A data frame with three columns (\code{item}, \code{type}, and \code{citation})
containing the references.
}
\description{
Creates a bibliography from a table of detections, with references for:
\itemize{
\item The ETN data platform.
\item The etn R package.
\item Animal project(s) associated with the animals that were detected.
\item Acoustic project(s) associated with the receivers from which the detections
were obtained.
}

It is recommended to cite these when using the data, see the \href{https://europeantrackingnetwork.org/en/4-data-policy-permissions-citation-guidelines-and-data-use}{ETN Citation Guidelines}
for details.
}
\examples{
\dontshow{if (interactive() && etn:::credentials_are_set()) withAutoprint(\{ # examplesIf}
# Create a bibliography from queried detections
get_acoustic_detections(scientific_name = "Mola mola") |>
  get_bibliography()

# Or obtain the bibliography from a Data Package
read_resource(example_dataset(), "bibliography")
\dontshow{\}) # examplesIf}
}
\seealso{
Other access functions:
\code{\link[=get_acoustic_deployments]{get_acoustic_deployments()}},
\code{\link[=get_acoustic_detections]{get_acoustic_detections()}},
\code{\link[=get_acoustic_projects]{get_acoustic_projects()}},
\code{\link[=get_acoustic_receivers]{get_acoustic_receivers()}},
\code{\link[=get_animal_projects]{get_animal_projects()}},
\code{\link[=get_animals]{get_animals()}},
\code{\link[=get_cpod_projects]{get_cpod_projects()}},
\code{\link[=get_package]{get_package()}},
\code{\link[=get_tags]{get_tags()}}
}
\concept{access functions}
