% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bibliography.R
\name{get_bibliography}
\alias{get_bibliography}
\title{Generate a bibliography for a data.frame of detections}
\usage{
get_bibliography(x)
}
\arguments{
\item{x}{A data.frame containing at least the columns \code{animal_project_code}
and \code{acoustic_project_code}. For example a data.frame returned by
\code{get_detections()}.}
}
\value{
A data.frame with three columns: \code{item}, \code{type}, and \code{citation}. The
first row contains the citation for the ETN platform, the second row the
citation for the etn R package, and subsequent rows contain citations for
the animal and acoustic projects present in the input data.frame.
}
\description{
This function generates a bibliography for a data.frame containing
detections, including citations for the ETN platform, the etn R package, and
any animal and acoustic projects present in the input data.frame.
}
\examples{
\dontshow{if (interactive() && etn:::credentials_are_set()) withAutoprint(\{ # examplesIf}
# Create a bibliography for a data.frame created by get_detections()
my_detections <-
  get_acoustic_detections(scientific_name = "Mola mola")
  get_bibliography(my_detections)

# Or from a frictionless datapackage containing detections
read_resource(example_dataset(), "detections") |> get_bibliography()

# Altough mainly meant for the above use cases, you can also provide any
# data.frame as long as the project code columns are present.
data.frame(
  animal_project_code = "2014_demer", acoustic_project_code = "demer" ) |>
  get_bibliography()
\dontshow{\}) # examplesIf}
}
\seealso{
Other access functions:
\code{\link[=get_acoustic_deployments]{get_acoustic_deployments()}},
\code{\link[=get_acoustic_detections]{get_acoustic_detections()}},
\code{\link[=get_acoustic_projects]{get_acoustic_projects()}},
\code{\link[=get_acoustic_receivers]{get_acoustic_receivers()}},
\code{\link[=get_animal_projects]{get_animal_projects()}},
\code{\link[=get_animals]{get_animals()}},
\code{\link[=get_cpod_projects]{get_cpod_projects()}},
\code{\link[=get_package]{get_package()}},
\code{\link[=get_tags]{get_tags()}}
}
\concept{access functions}
