% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpod_projects.R
\name{get_cpod_projects}
\alias{get_cpod_projects}
\title{Get cpod project data}
\usage{
get_cpod_projects(connection, cpod_project_code = NULL, citation = FALSE)
}
\arguments{
\item{connection}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A connection to the ETN
database. This argument is no longer used. You will be prompted for
credentials instead.}

\item{cpod_project_code}{Character (vector). One or more cpod project
codes. Case-insensitive.}

\item{citation}{Logical. If \code{TRUE}, add citation information from IMIS/MarineInfo for each
project, based on the \code{imis_dataset_id} field. This will add the following columns to the output:
\itemize{
\item \code{citation}: A formatted citation with DOI if available.
\item \code{doi}: The DOI for the dataset if available.
\item \code{contact_name}: The contact person, usually the first author. If no contact person is entered, the first author with status creator.
\item \code{contact_email}: The email of the contact person.
\item \code{contact_affiliation}: The institute of the contact person.
}}
}
\value{
A tibble with animal project data, sorted by \code{project_code}.
}
\description{
Get data for cpod projects, with options to filter results.
}
\examples{
\dontshow{if (etn:::credentials_are_set()) withAutoprint(\{ # examplesIf}
# Get all animal projects
get_cpod_projects()

# Get a specific animal project
get_cpod_projects(cpod_project_code = "cpod-lifewatch")
\dontshow{\}) # examplesIf}
}
