% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpod_projects.R
\name{get_cpod_projects}
\alias{get_cpod_projects}
\title{Get cpod project data}
\usage{
get_cpod_projects(connection, cpod_project_code = NULL, api = TRUE)
}
\arguments{
\item{connection}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A connection to the ETN
database. This argument is no longer used. You will be prompted for
credentials instead.}

\item{cpod_project_code}{Character (vector). One or more cpod project
codes. Case-insensitive.}

\item{api}{Logical. If \code{TRUE}, the function will use the API to access ETN.}
}
\value{
A tibble with animal project data, sorted by \code{project_code}. See
also
\href{https://inbo.github.io/etn/articles/etn_fields.html}{field definitions}.
}
\description{
Get data for cpod projects, with options to filter results.
}
\examples{
# Get all animal projects
get_cpod_projects()

# Get a specific animal project
get_cpod_projects(cpod_project_code = "cpod-lifewatch")
}
