% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpod_projects.R
\name{get_cpod_projects}
\alias{get_cpod_projects}
\title{Get cpod project data}
\usage{
get_cpod_projects(connection = con, cpod_project_code = NULL)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}

\item{cpod_project_code}{Character (vector). One or more cpod project
codes. Case-insensitive.}
}
\value{
A tibble with animal project data, sorted by \code{project_code}. See
also
\href{https://inbo.github.io/etn/articles/etn_fields.html}{field definitions}.
}
\description{
Get data for cpod projects, with options to filter results.
}
\examples{
# Set default connection variable
con <- connect_to_etn()

# Get all animal projects
get_cpod_projects(con)

# Get a specific animal project
get_cpod_projects(con, cpod_project_code = "cpod-lifewatch")
}
