% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deployments.R
\name{get_deployments}
\alias{get_deployments}
\title{Get deployments data}
\usage{
get_deployments(connection, network_project = NULL, receiver_status = NULL)
}
\arguments{
\item{connection}{A valid connection with the ETN database.}

\item{network_project}{(string) One or more network projects.}

\item{receiver_status}{(string) One or more receiver status.}
}
\value{
A tibble (tidyverse data.frame).
}
\description{
Get all or specific, filtered by network project and/or the status of the
receiver, deployments data.
}
\examples{
\dontrun{
con <- connect_to_etn(your_username, your_password)

# All deployments
get_deployments(con)

# Deployments of a subset of projects
get_deployments(con, network_project = c("zeeschelde", "ws1"))

# Deployments of a subset of projects and receiver status
get_deployments(con, network_project = c("zeeschelde", "ws1"),
                receiver_status = "Active")

# Deployments of a subset of receiver status
get_deployments(con, receiver_status = c("Broken", "Lost"))
}
}
