% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deployments.R
\name{get_deployments}
\alias{get_deployments}
\title{Get deployments data}
\usage{
get_deployments(connection, network_project = NULL, receiver_status = NULL)
}
\arguments{
\item{receiver_status}{}
}
\value{
data.frame
}
\description{
...
}
\examples{
\\notrun{
# All deployments
get_deployments(con)

# Deployments subset of projects
get_deployments(con, network_project = c("zeeschelde", "ws1"))

# Deployments subset of projects and receiver status
get_deployments(con, network_project = c("zeeschelde", "ws1"),
                receiver_status = "Active")
# Deployments subset of receiver status
get_deployments(con, receiver_status = c("Broken", "Lost"))
}
}
