% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deployments.R
\name{get_deployments}
\alias{get_deployments}
\title{Get deployment data}
\usage{
get_deployments(
  connection = con,
  application_type = NULL,
  network_project_code = NULL,
  open_only = TRUE
)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}

\item{application_type}{Character. \code{acoustic_telemetry} or \code{cpod}.}

\item{network_project_code}{Character (vector). One or more network
projects.}

\item{receiver_status}{Character (vector). One or more receiver status.}

\item{open_only}{Logical. Restrict deployments to those that are currently
open (i.e. no end date defined). Defaults to \code{TRUE}.}
}
\value{
A tibble with deployments data, sorted by \code{network_project_code},
\code{station_name} and \code{deploy_date_time}. See also
\href{https://inbo.github.io/etn/articles/etn_fields.html}{field definitions}.
}
\description{
Get data for deployments, with options to filter on network project,
receiver status and/or open deployments.
}
\examples{
\dontrun{
con <- connect_to_etn(your_username, your_password)

# Get all open deployments
get_deployments(con)

# Get all deployments (including those with an end date)
get_deployments(con, open_only = FALSE)

# Get open deployments for a specific application type
get_deployments(con, application_type = "cpod")

# Get open deployments from specific network project(s)
get_deployments(con, network_project_code = c("ws1", "ws2"))
}
}
