% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deployments.R
\name{get_deployments}
\alias{get_deployments}
\title{Get deployment metadata}
\usage{
get_deployments(
  connection = con,
  application_type = NULL,
  network_project_code = NULL,
  receiver_status = NULL,
  open_only = TRUE
)
}
\arguments{
\item{connection}{A valid connection to the ETN database.}

\item{application_type}{(string) \code{acoustic_telemetry} or \code{cpod}.}

\item{network_project_code}{(string) One or more network projects.}

\item{receiver_status}{(string) One or more receiver status.}

\item{open_only}{(logical) Restrict to deployments that are currently open
(i.e. no end date defined). Default: \code{TRUE}.}
}
\value{
A tibble (tidyverse data.frame) with metadata for deployments,
sorted by \code{network_project_code}, \code{station_name} and \code{deploy_date_time}.
}
\description{
Get metadata for deployments, with options to filter on network project,
receiver status and/or open deployments.
}
\examples{
\dontrun{
con <- connect_to_etn(your_username, your_password)

# Get all open deployments
get_deployments(con)

# Get all deployments (including those with an end date)
get_deployments(con, open_only = FALSE)

# Get open deployments for a specific application type
get_deployments(con, application_type = "cpod")

# Get open deployments from specific network project(s)
get_deployments(con, network_project_code = c("ws1", "ws2"))

# Get open deployments with a specific receiver status
get_deployments(con, receiver_status = c("Broken", "Lost"))

# Get open deployments from a specific project from active receivers
get_deployments(con, network_project_code = "ws1", receiver_status = "Active")
}
}
