% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deployments.R
\name{get_deployments}
\alias{get_deployments}
\title{Get deployments data}
\usage{
get_deployments(connection, network_project = NULL,
  receiver_status = NULL, open_only = TRUE)
}
\arguments{
\item{connection}{A valid connection with the ETN database.}

\item{network_project}{(string) One or more network projects.}

\item{receiver_status}{(string) One or more receiver status.}

\item{open_only}{(logical) Default TRUE, returning only those deployments
that are currently open (i.e. no end date defined). If FALSE, all deployments
are returned.}
}
\value{
A tibble (tidyverse data.frame).
}
\description{
Get all or specific, filtered by network project and/or the status of the
receiver, deployments data.
}
\examples{
\dontrun{
con <- connect_to_etn(your_username, your_password)

# All open deployments
get_deployments(con)

# All deployments (also deployments with an end date)
get_deployments(con, open_only = FALSE)

# Open deployments of a subset of projects
get_deployments(con, network_project = c("thornton", "leopold"))

# Open deployments of a subset of receiver status
get_deployments(con, receiver_status = c("Broken", "Lost"))

# Open deployments of a subset of projects and receiver status
get_deployments(con, network_project = "thornton",
                receiver_status = "Active")
}
}
