% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package.R
\name{get_package}
\alias{get_package}
\title{Get animal project data as a Data Package}
\usage{
get_package(animal_project_code)
}
\arguments{
\item{animal_project_code}{Animal project you want to get data from.}
}
\value{
A Data Package object.
}
\description{
Gets data related to an \strong{animal project} as a \href{https://specs.frictionlessdata.io/data-package/}{Data Package}.
}
\section{Included resources}{

The Data Package will contain and describe the following resources:
\itemize{
\item \code{animals}: Animals related to an \code{animal_project_code}, as returned by
\code{\link[=get_animals]{get_animals()}}.
\item \code{tags}: Tags associated with the selected animals, as returned by
\code{\link[=get_tags]{get_tags()}}.
\item \code{detections}: Acoustic detections for the selected animals, as returned by
\code{\link[=get_acoustic_detections]{get_acoustic_detections()}}.
\item \code{deployments}: Acoustic deployments for the \code{acoustic_project_code}(s)
found in detections, as returned by \code{\link[=get_acoustic_deployments]{get_acoustic_deployments()}}.
This allows you to see when receivers were deployed, even if these did
not detect the selected animals.
\item \code{receivers}: Acoustic receivers for the selected deployments, as returned
by \code{\link[=get_acoustic_receivers]{get_acoustic_receivers()}}.
}

You can write the Data Package to disk with \code{\link[=write_package]{write_package()}}.
}

\section{Data quality}{

The data are downloaded from the ETN database \emph{as is}, i.e. no quality or
consistency checks are performed.
We therefore recommend to verify the data before publication.
A consistency check can be performed on the files using:

\if{html}{\out{<div class="sourceCode">}}\preformatted{pip install frictionless
frictionless validate datapackage.json
}\if{html}{\out{</div>}}
}

\examples{
\dontshow{if (etn:::credentials_are_set() & interactive()) withAutoprint(\{ # examplesIf}
# Get a Data Package for a project
get_package(animal_project_code = "2014_demer")
\dontshow{\}) # examplesIf}
}
\seealso{
Other access functions: 
\code{\link{get_acoustic_deployments}()},
\code{\link{get_acoustic_detections}()},
\code{\link{get_acoustic_projects}()},
\code{\link{get_acoustic_receivers}()},
\code{\link{get_animal_projects}()},
\code{\link{get_animals}()},
\code{\link{get_cpod_projects}()},
\code{\link{get_tags}()}
}
\concept{access functions}
