% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package.R
\name{get_package}
\alias{get_package}
\title{Get acoustic data package}
\usage{
get_package(animal_project_code)
}
\arguments{
\item{animal_project_code}{Character. Animal project you want to get an
acoustic data package for. Required.}
}
\value{
A Data Package object.
}
\description{
Get all acoustic data related to an \strong{animal project}  as a data package.
}
\details{
Get all acoustic data related to an \strong{animal project}  as a
\strong{\href{https://specs.frictionlessdata.io/data-package/}{Frictionless Data Package}}
containing:\tabular{ll}{
   \strong{file} \tab \strong{description} \cr
   \code{animals.csv} \tab Animals related to an \code{animal_project_code}, as returned by \code{get_animals()}. \cr
   \code{tags.csv} \tab Tags associated with the selected animals, as returned by \code{get_tags()}. \cr
   \code{detections.csv} \tab Acoustic detections for the selected animals, as returned by \code{get_acoustic_detections()}. \cr
   \code{deployments.csv} \tab Acoustic deployments for the \code{acoustic_project_code}(s) found in detections, as returned by \code{get_acoustic_deployments()}. This allows users to see when receivers were deployed, even if these did not detect the selected animals. \cr
   \code{receivers.csv} \tab Acoustic receivers for the selected deployments, as returned by \code{get_acoustic_receivers()}. \cr
   \code{datapackage.json} \tab A \href{https://specs.frictionlessdata.io/table-schema/}{Frictionless Table Schema} metadata file describing the fields and relations of the above csv files. \cr
}


\strong{Important}: The data are downloaded \emph{as is} from the database, i.e. no
quality or consistency checks are performed by this function. We therefore
recommend to verify the data before publication. A consistency check can be
performed by validation tools of the Frictionless Framework, e.g.
\verb{frictionless validate datapackage.json} on the command line using
\href{https://github.com/frictionlessdata/frictionless-py}{frictionless-py}.
}
\examples{
\dontshow{if (etn:::credentials_are_set() & interactive()) withAutoprint(\{ # examplesIf}
# Get data package for the 2014_demer animal project
get_package(animal_project_code = "2014_demer")
#> A Data Package with 5 resources:
#> • animals
#> • tags
#> • detections
#> • deployments
#> • receivers
#> Use `unclass()` to print the Data Package as a list.
\dontshow{\}) # examplesIf}
}
\seealso{
Other access functions: 
\code{\link{get_acoustic_deployments}()},
\code{\link{get_acoustic_detections}()},
\code{\link{get_acoustic_projects}()},
\code{\link{get_acoustic_receivers}()},
\code{\link{get_animal_projects}()},
\code{\link{get_animals}()},
\code{\link{get_cpod_projects}()},
\code{\link{get_tags}()}
}
\concept{access functions}
