% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projects.R
\name{get_projects}
\alias{get_projects}
\title{Get project overview}
\usage{
get_projects(connection, project_type = NULL)
}
\arguments{
\item{project_type}{animal | network}
}
\value{
data.frame
}
\description{
When no project_type ...
}
