% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projects.R
\name{get_projects}
\alias{get_projects}
\title{Get project metadata}
\usage{
get_projects(connection, project_type = NULL)
}
\arguments{
\item{connection}{A valid connection to the ETN database.}

\item{project_type}{(string) Either \code{animal} or \code{network}.}
}
\value{
A tibble (tidyverse data.frame).
}
\description{
Get metadata for projects, with option to filter on animal or network
projects.
}
\examples{
\dontrun{
con <- connect_to_etn(your_username, your_password)

# Get all projects
get_projects(con)

# Get all animal projects
get_projects(con, project_type = "animal")

# Get all network projects
get_projects(con, project_type = "network")
}
}
