% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projects.R
\name{get_projects}
\alias{get_projects}
\title{Get project overview}
\usage{
get_projects(connection, project_type = NULL)
}
\arguments{
\item{connection}{connection to the ETN database}

\item{project_type}{animal | network}
}
\value{
data.frame
}
\description{
Get an overview of the
}
\examples{
\dontrun(
get_projects()
get_projects(project_type = "animal")
)
}
