% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projects.R
\name{get_projects}
\alias{get_projects}
\title{Get project overview}
\usage{
get_projects(connection, project_type = NULL)
}
\arguments{
\item{connection}{A valid connection with the ETN database.}

\item{project_type}{(string) Either \code{animal} or \code{network}.}
}
\value{
A data.frame.
}
\description{
Get an overview of the projects available on ETN database.
}
\examples{
\dontrun{
# Get a list of all projects
get_projects()

# Get a list of all animal projects
get_projects(project_type = "animal")

# Get a list of all network projects
get_projects(project_type = "network")
}
}
