% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_receiver_diagnostics.R
\name{get_receiver_diagnostics}
\alias{get_receiver_diagnostics}
\title{Get diagnostics information for a receiver}
\usage{
get_receiver_diagnostics(connection = con, limit = FALSE)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}

\item{limit}{Logical. Limit the number of returned records to 100 (useful
for testing purposes). Defaults to \code{FALSE}.}
}
\value{
A tibble with receiver diagnostics data
}
\description{
Get diagnostics information for a receiver
}
