% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_receiver_diagnostics.R
\name{get_receiver_diagnostics}
\alias{get_receiver_diagnostics}
\title{Get diagnostics information for a receiver}
\usage{
get_receiver_diagnostics(
  connection = con,
  start_date = NULL,
  end_date = NULL,
  receiver_id = NULL,
  deployment_id = NULL,
  limit = FALSE
)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}

\item{start_date}{Character. Start date (inclusive) in ISO 8601 format (
\code{yyyy-mm-dd}, \code{yyyy-mm} or \code{yyyy}).}

\item{end_date}{Character. End date (exclusive) in ISO 8601 format (
\code{yyyy-mm-dd}, \code{yyyy-mm} or \code{yyyy}).}

\item{receiver_id}{Character (vector). One or more receiver identifiers.}

\item{deployment_id}{Integer (vector). One or more deployment identifiers.}

\item{limit}{Logical. Limit the number of returned records to 100 (useful
for testing purposes). Defaults to \code{FALSE}.}
}
\value{
A tibble with receiver diagnostics data
}
\description{
Get diagnostics information for a receiver
}
