% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_receivers.R
\name{get_receivers}
\alias{get_receivers}
\title{Get receiver metadata}
\usage{
get_receivers(connection = con)
}
\arguments{
\item{connection}{A valid connection to the ETN database.}

\item{receiver_id}{(string) One or more receiver ids.}
}
\value{
A tibble (tidyverse data.frame).
}
\description{
Get metadata for receivers.
}
\examples{
\dontrun{
con <- connect_to_etn(your_username, your_password)

# Get all receivers
get_receivers(con)
}
}
