% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_receivers.R
\name{get_receivers}
\alias{get_receivers}
\title{Get receiver data}
\usage{
get_receivers(connection, network_project = NULL)
}
\arguments{
\item{connection}{A valid connection with the ETN database.}

\item{network_project}{(string) One or more network projects.}
}
\value{
A tibble (tidyverse data.frame).
}
\description{
Get all or specific, filtered on network project, receiver data.
}
\examples{
\dontrun{
con <- connect_to_etn(your_username, your_password)

# Get all receivers data
get_receivers(con)

# Get receivers data filtered on a single network project
get_receivers(con, network_project = "demer")

# Get receivers data filtered on a multiple network projects
get_receivers(con, network_project = c("demer", "dijle"))
}
}
