% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_receivers.R
\name{get_receivers}
\alias{get_receivers}
\title{Get receiver metadata}
\usage{
get_receivers(connection, network_project = NULL)
}
\arguments{
\item{connection}{A valid connection with the ETN database.}

\item{network_project}{(string) One or more network projects.}
}
\value{
A tibble (tidyverse data.frame).
}
\description{
Get metadata for receivers, with option to filter on network project.
}
\examples{
\dontrun{
con <- connect_to_etn(your_username, your_password)

# Get metadata for all receivers
receivers <- get_receivers(con)

# Get metadata for receivers linked to specific network project(s)
receivers <- get_receivers(con, network_project = "thornton")
receivers <- get_receivers(con, network_project = c("thornton", "2012_leopoldkanaal"))
}
}
