% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tags.R
\name{get_tags}
\alias{get_tags}
\title{Get tag data}
\usage{
get_tags(
  connection,
  tag_type = NULL,
  tag_subtype = NULL,
  tag_serial_number = NULL,
  acoustic_tag_id = NULL
)
}
\arguments{
\item{connection}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A connection to the ETN
database. This argument is no longer used. You will be prompted for
credentials instead.}

\item{tag_type}{Character (vector). \code{acoustic} or \code{archival}. Some tags are
both, find those with \code{acoustic-archival}.}

\item{tag_subtype}{Character (vector). \code{animal}, \verb{built-in}, \code{range} or
\code{sentinel}.}

\item{tag_serial_number}{Character (vector). One or more tag serial numbers.}

\item{acoustic_tag_id}{Character (vector). One or more acoustic tag
identifiers, i.e. identifiers found in \code{\link[=get_acoustic_detections]{get_acoustic_detections()}}.}
}
\value{
A tibble with tags data, sorted by \code{tag_serial_number}.
Values for \code{owner_organization} and \code{owner_pi} will only be visible if you
are member of the group.
}
\description{
Get data for tags, with options to filter results. Note that there
can be multiple records (\code{acoustic_tag_id}) per tag device
(\code{tag_serial_number}).
}
\examples{
\dontshow{if (etn:::credentials_are_set()) withAutoprint(\{ # examplesIf}
# Get all tags
get_tags()

# Get archival tags, including acoustic-archival
get_tags(tag_type = c("archival", "acoustic-archival"))

# Get tags of specific subtype
get_tags(tag_subtype = c("built-in", "range"))

# Get specific tags (note that these can return multiple records)
get_tags(tag_serial_number = "1187450")
get_tags(acoustic_tag_id = "A69-1601-16130")
get_tags(acoustic_tag_id = c("A69-1601-16129", "A69-1601-16130"))
\dontshow{\}) # examplesIf}
}
\seealso{
Other access functions:
\code{\link[=get_acoustic_deployment_logs]{get_acoustic_deployment_logs()}},
\code{\link[=get_acoustic_deployments]{get_acoustic_deployments()}},
\code{\link[=get_acoustic_detections]{get_acoustic_detections()}},
\code{\link[=get_acoustic_projects]{get_acoustic_projects()}},
\code{\link[=get_acoustic_receivers]{get_acoustic_receivers()}},
\code{\link[=get_animal_projects]{get_animal_projects()}},
\code{\link[=get_animals]{get_animals()}},
\code{\link[=get_bibliography]{get_bibliography()}},
\code{\link[=get_cpod_projects]{get_cpod_projects()}},
\code{\link[=get_package]{get_package()}}
}
\concept{access functions}
