% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tags.R
\name{get_tags}
\alias{get_tags}
\title{Get tag data}
\usage{
get_tags(connection = con, tag_serial_number = NULL, acoustic_tag_id = NULL)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}

\item{tag_serial_number}{Character (vector). One or more tag serial numbers.}

\item{acoustic_tag_id}{Character (vector). One or more acoustic tag
identifiers (i.e. identifier for detections).}

\item{tag_type}{Character (vector).}

\item{tag_subtype}{Character (vector).}
}
\value{
A tibble with tags data, sorted by \code{tag_serial_number}. See also
\href{https://inbo.github.io/etn/articles/etn_fields.html}{field definitions}.
}
\description{
Get data for tags, with options to filter results. Note that there
can be multiple records (\code{acoustic_tag_id}) per tag device
(\code{tag_serial_number}).
}
\examples{
\dontrun{
# Set default connection variable
con <- connect_to_etn()

# Get all tags
get_tags()

# Get specific tags
get_tags(tag_serial_number = "1187450")
get_tags(acoustic_tag_id = "A69-1601-16130")
get_tags(acoustic_tag_id = c("A69-1601-16129", "A69-1601-16130"))
}
}
