% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tags.R
\name{get_tags}
\alias{get_tags}
\title{Get tag metadata}
\usage{
get_tags(connection, animal_project = NULL, include_reference_tags = FALSE)
}
\arguments{
\item{connection}{A valid connection with the ETN database.}

\item{animal_project}{(string) One or more animal projects.}

\item{include_reference_tags}{(logical) Include (non-animal borne) reference
tags. Default: FALSE.}
}
\value{
A tibble (tidyverse data.frame).
}
\description{
Get metadata of tags. Only tags that can be linked to an animal (and thus an
\code{animal_project_code}) are returned to the user. By default, reference tags
are excluded.
}
\examples{
\dontrun{
con <- connect_to_etn(your_username, your_password)

# Get metadata of all (animal) tags
tags <- get_tags(con)

# Get metadata of all tags, including reference tags
tags <- get_tags(con, include_reference_tags = TRUE)

# Get metadata of tags linked to specific project(s)
tags <- get_tags(con, animal_project = "phd_reubens")
tags <- get_tags(con, animal_project = c("phd_reubens", "2012_leopoldkanaal"))
}
}
