% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tags.R
\name{get_tags}
\alias{get_tags}
\title{Get tag metadata}
\usage{
get_tags(connection = con, tag_id = NULL, include_ref_tags = FALSE)
}
\arguments{
\item{connection}{A valid connection with the ETN database.}

\item{tag_id}{(string) One or more tag ids.}

\item{include_ref_tags}{(logical) Include reference tags. Default:
\code{FALSE}.}
}
\value{
A tibble (tidyverse data.frame).
}
\description{
Get metadata for tags, with option to filter on tag id. By default, reference
tags are excluded.
}
\examples{
\dontrun{
con <- connect_to_etn(your_username, your_password)

# Get all (animal) tags
get_tags(con)

# Get all tags, including reference tags
get_tags(con, include_ref_tags = TRUE)

# Get specific tags (will automatically set include_ref_tags = TRUE)
get_tags(con, tag_id = "A69-1303-65313")
get_tags(con, tag_id = c("A69-1601-1705", "A69-1601-1707"))
}
}
