% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tags.R
\name{get_tags}
\alias{get_tags}
\title{Get tag data}
\usage{
get_tags(connection = con, tag_id = NULL, include_ref_tags = FALSE)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}

\item{tag_id}{Character (vector). One or more tag ids.}

\item{include_ref_tags}{Logical. Include reference tags. Defaults to
\code{FALSE}.}
}
\value{
A tibble with tags data, sorted by \code{tag_id}. See also
\href{https://inbo.github.io/etn/articles/etn_fields.html}{field definitions}.
}
\description{
Get data for tags, with options to filter results. By default, reference
tags are excluded.
}
\examples{
\dontrun{
con <- connect_to_etn(your_username, your_password)

# Get all (animal) tags
get_tags(con)

# Get all tags, including reference tags
get_tags(con, include_ref_tags = TRUE)

# Get specific tags (will automatically set include_ref_tags = TRUE)
get_tags(con, tag_id = "A69-1303-65313")
get_tags(con, tag_id = c("A69-1601-1705", "A69-1601-1707"))
}
}
