% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tags.R
\name{get_tags}
\alias{get_tags}
\title{Get tags metadata}
\usage{
get_tags(connection, animal_project = NULL)
}
\arguments{
\item{connection}{A valid connection with the ETN database.}

\item{animal_project}{(string) One or more animal projects.}
}
\value{
A data.frame.
}
\description{
Get the metadate about the transmitter tags. At the moment, only tags that
can be linked to a projectcode are returned to the user.
}
\examples{
\dontrun{
# Get the metadata of all transmitter tags
get_tags(con)

# Get the metadata of the tags linked to specific project(s)
get_tags(con, animal_project = "phd_reubens")
get_tags(con, animal_project = c("phd_reubens", "2013_albertkanaal"))
}
}
