% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tags.R
\name{get_tags}
\alias{get_tags}
\title{Get tag metadata}
\usage{
get_tags(
  connection = con,
  animal_project_code = NULL,
  include_reference_tags = FALSE
)
}
\arguments{
\item{connection}{A valid connection with the ETN database.}

\item{animal_project_code}{(string) One or more animal projects.}

\item{include_reference_tags}{(logical) Include reference tags. Default:
\code{FALSE}.}
}
\value{
A tibble (tidyverse data.frame).
}
\description{
Get metadata for tags. Only returns tags that can be linked to an animal (and
thus an animal project). By default, reference tags are excluded.
}
\examples{
\dontrun{
con <- connect_to_etn(your_username, your_password)

# Get all (animal) tags
get_tags(con)

# Get all tags, including reference tags
get_tags(con, include_reference_tags = TRUE)

# Get tags linked to specific animal project(s)
get_tags(con, animal_project_code = "phd_reubens")
get_tags(con, animal_project_code = c("phd_reubens", "2012_leopoldkanaal"))
}
}
