% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_animal_ids_api.R
\name{get_val}
\alias{get_val}
\title{Retrieve the result of a function called to the opencpu api}
\usage{
get_val(temp_key, api_domain = "https://opencpu.lifewatch.be")
}
\arguments{
\item{temp_key}{the temp key returned from the POST request to the API}
}
\value{
the uncompressed object resulting form a GET request to the API
}
\description{
Loading the evaluated object into the current environment
}
\examples{
etn:::extract_temp_key(response) \%>\% get_val()
}
