% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_acoustic_tag_ids.R
\name{list_acoustic_tag_ids}
\alias{list_acoustic_tag_ids}
\title{List all available acoustic tag ids}
\usage{
list_acoustic_tag_ids(connection = con)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}
}
\value{
A vector of all unique \code{acoustic_tag_id} in \code{acoustic_tag_id.sql}.
}
\description{
List all available acoustic tag ids
}
