% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_animal_project_codes.R
\name{list_animal_project_codes}
\alias{list_animal_project_codes}
\title{List all available animal project codes}
\usage{
list_animal_project_codes(api = TRUE, connection)
}
\arguments{
\item{api}{Logical. If \code{TRUE}, the function will use the API to access ETN.}

\item{connection}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A connection to the ETN
database. This argument is no longer used. You will be prompted for
credentials instead.}
}
\value{
A vector of all unique \code{project_code} of \code{type = "animal"} in
\code{project.sql}.
}
\description{
List all available animal project codes
}
