% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_network_project_codes.R
\name{list_network_project_codes}
\alias{list_network_project_codes}
\title{List all available network project codes}
\usage{
list_network_project_codes(connection = con)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}
}
\value{
A vector of all unique \code{project_code} (lowercased) of
\code{project_type="network"} present in \code{projects_view2}.
}
\description{
List all available network project codes
}
