% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_receiver_ids.R
\name{list_receiver_ids}
\alias{list_receiver_ids}
\title{List all available receiver ids}
\usage{
list_receiver_ids(api = TRUE, connection)
}
\arguments{
\item{api}{Logical. If \code{TRUE}, the function will use the API to access ETN.}

\item{connection}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A connection to the ETN
database. This argument is no longer used. You will be prompted for
credentials instead.}
}
\value{
A vector of all unique \code{receiver} present in \code{acoustic.receivers}.
}
\description{
List all available receiver ids
}
