% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_scientific_names.R
\name{list_scientific_names}
\alias{list_scientific_names}
\title{List all available scientific names}
\usage{
list_scientific_names(api = TRUE, connection)
}
\arguments{
\item{api}{Logical. If \code{TRUE}, the function will use the API to access ETN.}

\item{connection}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A connection to the ETN
database. This argument is no longer used. You will be prompted for
credentials instead.}
}
\value{
A vector of all unique \code{scientific_name} present in
\code{common.animal_release}.
}
\description{
List all available scientific names
}
