% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_scientific_names.R
\name{list_scientific_names}
\alias{list_scientific_names}
\title{List all available scientific names}
\usage{
list_scientific_names(connection = con)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}
}
\value{
A vector of all unique \code{scientific_name} present in
\code{common.animal_release}.
}
\description{
List all available scientific names
}
