% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_station_names.R
\name{list_station_names}
\alias{list_station_names}
\title{List all available station names}
\usage{
list_station_names(connection = con)
}
\arguments{
\item{connection}{A connection to the ETN database. Defaults to \code{con}.}
}
\value{
A vector of all unique \code{station_name} present in
\code{acoustic.deployments}.
}
\description{
List all available station names
}
