% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tag_serial_numbers.R
\name{list_tag_serial_numbers}
\alias{list_tag_serial_numbers}
\title{List all available tag serial numbers}
\usage{
list_tag_serial_numbers(connection, api = TRUE)
}
\arguments{
\item{connection}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A connection to the ETN
database. This argument is no longer used. You will be prompted for
credentials instead.}

\item{api}{Logical. If \code{TRUE}, the function will use the API to access ETN.}
}
\value{
A vector of all unique \code{tag_serial_numbers} present in
\code{common.tag_device}.
}
\description{
List all available tag serial numbers
}
