% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_values.R
\name{list_values}
\alias{list_values}
\title{Get all distinct values from a specific column of a data.frame}
\usage{
list_values(.data, column, split = ",")
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble)}

\item{column}{unquoted or quoted column name or column position}

\item{split}{character vector containing regular expression(s) to use for
splitting. \code{split} is internally passed to argument \code{split} of base R
function \code{strsplit()}. Default: \code{","} (comma-separated values)}
}
\value{
a vector of the same type as the given column
}
\description{
This function extracts all distinct values from a given column of a
data.frame. If the select column contains characters, multiple values in the
same row are split.
}
\examples{
\dontrun{
df <- data.frame(
  chr_col = c("A", "B,C", "C,A","D"),
  num_col = c(1,2,2,3),
  dot_sep_col = c("A", "B.C", "C.A","D"))
# split by comma  (default)
list_values(df, chr_col) # same as
list_values(df, chr_col, split = ",")
# access column by column position
list_values(df, 1)
# split by dot (regex expression: "\\.")
list_values(df, dot_sep_col, split = "\\\\.")
# non character column
list_values(df, num_col)
# list_values can be added to a pipe
df \%>\% list_values(chr_col)
}
}
