% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{receiver_ids}
\alias{receiver_ids}
\title{Get list of unique available receiver_id}
\usage{
receiver_ids(connection)
}
\arguments{
\item{connection}{A valid connection to the ETN database.}
}
\value{
A vector of all receiver_id present in vliz.receivers_view2.
}
\description{
Get list of unique available receiver_id
}
