% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_animals.R
\name{scientific_names}
\alias{scientific_names}
\title{Support function to get unique set of scientific names}
\usage{
scientific_names(connection)
}
\arguments{
\item{connection}{A valid connection to ETN database.}
}
\value{
A vector of all scientific names present in vliz.animals_view.
}
\description{
This function retrieves all unique scientific names
}
