% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{station_names}
\alias{station_names}
\title{Get list of unique available station_name}
\usage{
station_names(connection)
}
\arguments{
\item{connection}{A valid connection to the ETN database.}
}
\value{
A vector of all station_name present in vliz.deployments_view2.
}
\description{
Get list of unique available station_name
}
