% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{tag_ids}
\alias{tag_ids}
\title{Get a list of unique available tag_id}
\usage{
tag_ids(connection)
}
\arguments{
\item{connection}{A valid connection to the ETN database.}
}
\value{
A vector of all tag_id present in vliz.tags_view2.
}
\description{
Get a list of unique available tag_id
}
