% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dwc.R
\name{write_dwc}
\alias{write_dwc}
\title{Transform ETN data to Darwin Core}
\usage{
write_dwc(
  connection = con,
  animal_project_code,
  directory = animal_project_code,
  rights_holder = NULL,
  license = "CC-BY-4.0"
)
}
\arguments{
\item{connection}{Connection to the ETN database.}

\item{animal_project_code}{Animal project code.}

\item{directory}{Path to local directory to write file(s) to.
Existing files of the same name will be overwritten.}

\item{rights_holder}{Acronym of the organization owning or managing the
rights over the data.}

\item{license}{Identifier of the license under which the data will be
published.
\itemize{
\item \href{https://creativecommons.org/licenses/by/4.0/legalcode}{\code{CC-BY-4.0}}
(default).
\item \href{https://creativecommons.org/publicdomain/zero/1.0/legalcode}{\code{CC0-1.0}}.
}}
}
\value{
CSV file(s) written to disk.
}
\description{
Transforms and downloads data from a European Tracking Network
\strong{animal project} to \href{https://dwc.tdwg.org/}{Darwin Core}.
The resulting CSV file(s) can be uploaded to a
\href{https://www.gbif.org/ipt}{GBIF IPT} for publication to OBIS or GBIF.
A \code{meta.xml} or \code{eml.xml} file are not created.
}
