% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dwc.R
\name{write_dwc}
\alias{write_dwc}
\title{Transform ETN data to Darwin Core}
\usage{
write_dwc(
  connection = con,
  animal_project_code,
  directory = ".",
  rights_holder = NULL,
  license = "CC-BY"
)
}
\arguments{
\item{connection}{Connection to the ETN database.}

\item{animal_project_code}{Animal project code.}

\item{directory}{Path to local directory to write file(s) to.
If \code{NULL}, then a list of data frames is returned instead, which can be
useful for extending/adapting the Darwin Core mapping before writing with
\code{\link[readr:write_delim]{readr::write_csv()}}.}

\item{rights_holder}{Acronym of the organization owning or managing the
rights over the data.}

\item{license}{Identifier of the license under which the data will be
published.
\itemize{
\item \href{https://creativecommons.org/licenses/by/4.0/legalcode}{\code{CC-BY}} (default).
\item \href{https://creativecommons.org/publicdomain/zero/1.0/legalcode}{\code{CC0}}.
}}
}
\value{
CSV file(s) written to disk or list of data frames when
\code{directory = NULL}.
}
\description{
Transforms and downloads data from a European Tracking Network
\strong{animal project} to \href{https://dwc.tdwg.org/}{Darwin Core}.
The resulting CSV file(s) can be uploaded to an \href{https://www.gbif.org/ipt}{IPT} for publication to OBIS and/or GBIF.
A \code{meta.xml} or \code{eml.xml} file are not created.
}
\section{Transformation details}{

Data are transformed into an
\href{https://rs.gbif.org/core/dwc_occurrence_2022-02-02.xml}{Occurrence core}.
This \strong{follows recommendations} discussed and created by Peter Desmet,
Jonas Mortelmans, Jonathan Pye, John Wieczorek and others.
See the \href{https://github.com/inbo/etn/tree/main/inst/sql}{SQL file(s)}
used by this function for details.

Key features of the Darwin Core transformation:
\itemize{
\item Deployments (animal+tag associations) are parent events, with capture,
surgery, release, recapture (human observations) and acoustic detections
(machine observations) as child events.
No information about the parent event is provided other than its ID,
meaning that data can be expressed in an Occurrence Core with one row per
observation and \code{parentEventID} shared by all occurrences in a deployment.
\item The release event often contains metadata about the animal (sex,
lifestage, comments) and deployment as a whole.
\item Acoustic detections are downsampled to the \strong{first detection per hour},
to reduce the size of high-frequency data.
Duplicate detections (same animal, tag and timestamp) are excluded.
It is possible for a deployment to contain no detections, e.g. if the
tag malfunctioned right after deployment.
}
}

