% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frictionless-reexports.R
\name{write_package}
\alias{write_package}
\title{Write a Data Package to disk}
\usage{
write_package(package, directory, compress = FALSE)
}
\description{
Write a Data Package to disk
}
\seealso{
Other frictionless functions: 
\code{\link{add_resource}()},
\code{\link{read_package}()}
}
\concept{frictionless functions}
